% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_units.R
\name{append_units.default}
\alias{append_units.default}
\title{Append Units By Default}
\usage{
\method{append_units}{default}(
  x,
  open = getOption("append_units_open", " ("),
  close = getOption("append_units_close", ")"),
  style = getOption("append_units_style", "character"),
  ...
)
}
\arguments{
\item{x}{object}

\item{open}{character to precede units}

\item{close}{character to follow units}

\item{style}{one of character, latex, or plotmath}

\item{...}{ignored arguments}
}
\value{
same class as x with named character label of length four having supplied class
}
\description{
Units attribute is wrapped in \code{open} and
\code{close}, and appended to label.
Label is then coerced using \code{as}
and concatenated using \code{\link{concatenate}}.
}
\examples{
library(units)
library(magrittr)
x <- 1:10
attr(x, 'label') <- 'acceleration'
units(x) <- 'm/s^2'
y <- as_units('kg')
x \%>\% attr('label')
x \%>\% append_units \%>\% attr('label')
y \%>\% attr('label')
y \%>\% append_units \%>\% attr('label')


}
\seealso{
Other labels: 
\code{\link{alias.data.frame}()},
\code{\link{append_units.data.frame}()},
\code{\link{append_units}()}
}
\concept{labels}
\keyword{internal}
