% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_units.R
\name{append_units}
\alias{append_units}
\title{Append Units}
\usage{
append_units(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Appends units attribute to label attribute.
Generic, with methods
\code{\link{append_units.default}} and
\code{\link{append_units.data.frame}}.
}
\examples{
# see methods
}
\seealso{
Other labels: 
\code{\link{alias.data.frame}()},
\code{\link{append_units.data.frame}()},
\code{\link{append_units.default}()}
}
\concept{labels}
\keyword{internal}
