% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{arbitrate.default}
\alias{arbitrate.default}
\title{Arbitrate Default}
\usage{
\method{arbitrate}{default}(x, y, tag = "", ...)
}
\arguments{
\item{x}{left attribute}

\item{y}{right attribute}

\item{...}{passed arguments}
}
\value{
list if y is list, else x
}
\description{
Arbitrates two attributes, the first of which is non-NULL and non-list.
If y is list, x is promoted to list and re-arbitrated thus.
Otherwise, x is returned, with warning if y not identical and not NULL.
}
\keyword{internal}
