% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{arbitrate.list}
\alias{arbitrate.list}
\title{Arbitrate List}
\usage{
\method{arbitrate}{list}(x, y, ...)
}
\arguments{
\item{x}{left attribute}

\item{y}{right attribute}

\item{...}{passed arguments}
}
\value{
list
}
\description{
Arbitrates two attributes, the first of which is a list.
If x has names, it is cast as namedList and arbitrated thus.
}
\keyword{internal}
