% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{arbitrate.namedList}
\alias{arbitrate.namedList}
\title{Arbitrate Named List}
\usage{
\method{arbitrate}{namedList}(x, y, ...)
}
\arguments{
\item{x}{left attribute}

\item{y}{right attribute}

\item{...}{passed arguments}
}
\value{
list
}
\description{
Arbitrates two attributes, the first of which is a named list.
}
\keyword{internal}
