% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{arbitrate.NULL}
\alias{arbitrate.NULL}
\title{Arbitrate Null}
\usage{
\method{arbitrate}{`NULL`}(x, y, ...)
}
\arguments{
\item{x}{left attribute}

\item{y}{right attribute}

\item{...}{passed arguments}
}
\value{
class of y
}
\description{
Arbitrates two attributes, the first of which is NULL.
Simply returns the second.
}
\keyword{internal}
