% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{arbitrate}
\alias{arbitrate}
\title{Arbitrate Two Attributes}
\usage{
arbitrate(x, y, ...)
}
\arguments{
\item{x}{left attribute set}

\item{y}{right attribute set}

\item{...}{passed arguments}
}
\value{
list (of attributes)
}
\description{
Arbitrates two attribute sets.  Generic, with methods
\code{\link{arbitrate.NULL}},
\code{\link{arbitrate.namedList}},
\code{\link{arbitrate.list}},
\code{\link{arbitrate.default}}.
}
\keyword{internal}
