% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{as.character.yam}
\alias{as.character.yam}
\title{Coerce Yam to Character}
\usage{
\method{as.character}{yam}(x, ...)
}
\arguments{
\item{x}{yam}

\item{...}{ignored; keys is an attribute of yam}
}
\value{
character
}
\description{
Coerces class yam to character.  Forms the basis for a
yamlet emitter.
}
\examples{
foo <- as_yamlet(c('id: subject','amt: dose'))
class(foo)
bar <- as_yam(foo)
class(bar)
as.character(bar)
as.character(
as_yam(
as_yamlet(
"race: [label: race, guide: [ white: 0, black: 1, asian: 2 ], multiple: ['yes': 1, 'no': 0]]"
)))

}
\seealso{
Other yam: 
\code{\link{as_yam.character}()},
\code{\link{as_yam.yamlet}()},
\code{\link{as_yam}()}
}
\concept{yam}
\keyword{internal}
