% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{as.expression.plotmath}
\alias{as.expression.plotmath}
\title{Coerce Plotmath to Expression}
\usage{
\method{as.expression}{plotmath}(x, ...)
}
\arguments{
\item{x}{plotmath}

\item{...}{ignored arguments}
}
\value{
expression, or list of expression for length(x) > 1
}
\description{
Coerces plotath to expression by parsing as text.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_wikisymbol(x)
x <- as_plotmath(x)
x
as.expression(x)[[4]]
as.expression(x[[4]])
class(as.expression(x))
lapply(as.expression(x), class)
as.expression(as_plotmath(as_wikisymbol('V_c./F')))
as.expression(as_plotmath(as_wikisymbol(character(0))))
library(magrittr)
'gravitational force (kg\\\\.m/s^2.)' \%>\%
  as_wikisymbol \%>\%
  as_plotmath \%>\%
  as_expression -> label
  label

}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol}()},
\code{\link{wikisym2latex_}()},
\code{\link{wikisym2plotmath_}()}
}
\concept{wikisymbol}
\keyword{internal}
