% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-to-plotmath.R
\name{as.png.wikisymbol}
\alias{as.png.wikisymbol}
\title{Convert Wiki Symbol to PNG}
\usage{
\method{as.png}{wikisymbol}(
  x,
  filename = tempfile(),
  width = 3,
  height = 1,
  units = "in",
  res = 150,
  ...
)
}
\arguments{
\item{x}{wikisymbol; see \code{\link{as_wikisymbol}}}

\item{filename}{path for image file}

\item{width}{width}

\item{height}{height}

\item{units}{units}

\item{res}{resolution}

\item{...}{passed arguments}
}
\value{
invisible filepath
}
\description{
Converts wikisymbol to png.
}
\examples{
library(magrittr)
'Omega ~ joule^\\\\*. ~1 kg*m^2./s^2' \%>\% as_wikisymbol \%>\% as.png -> file
file
}
\keyword{internal}
