% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_categorical.R
\name{as_categorical}
\alias{as_categorical}
\title{Coerce to Categorical}
\usage{
as_categorical(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed}
}
\description{
Coerce to categorical. 
Generic, with method \code{\link{as_categorical.decorated}}.
}
\seealso{
Other decorated: 
\code{\link{[.decorated}()},
\code{\link{[<-.decorated}()},
\code{\link{[[.decorated}()},
\code{\link{[[<-.decorated}()},
\code{\link{as_categorical.decorated}()},
\code{\link{merge.decorated}()},
\code{\link{names<-.decorated}()}
}
\concept{decorated}
\keyword{internal}
