% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified.R
\name{as_classified}
\alias{as_classified}
\title{Coerce to Classified}
\usage{
as_classified(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Coerce something to classified.
Generic, with method for factor.
}
\examples{
example(as_classified.factor)
}
\seealso{
Other classified: 
\code{\link{[.classified}()},
\code{\link{as_classified.factor}()}
}
\concept{classified}
\keyword{internal}
