% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{as_decorated.default}
\alias{as_decorated.default}
\title{Coerce to Decorated by Default}
\usage{
\method{as_decorated}{default}(x, meta = "-", ...)
}
\arguments{
\item{x}{object}

\item{meta}{see \code{\link{decorate.list}}}

\item{...}{passed arguments}
}
\value{
decorated
}
\description{
Coerces to class 'decorated' by decorating (by default) with an empty list.
}
\examples{
class(Puromycin)
class(as_decorated(Puromycin))
}
\seealso{
Other decorate: 
\code{\link{as_decorated}()},
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{decorate.list}()},
\code{\link{decorate}()},
\code{\link{decorations.data.frame}()},
\code{\link{decorations}()},
\code{\link{redecorate}()}
}
\concept{decorate}
\keyword{internal}
