% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{as_dvec.complex}
\alias{as_dvec.complex}
\title{Coerce Complex to Decorated Vector}
\usage{
\method{as_dvec}{complex}(x, ...)
}
\arguments{
\item{x}{complex}

\item{...}{attributes to assign}
}
\value{
dvec
}
\description{
Coerces complex to decorated vector.  Assigns class 'dvec' and any named attributes in dots.
}
\examples{
as_dvec(c(complex(1), complex(2)))
}
\seealso{
Other dvec: 
\code{\link{[.dvec}()},
\code{\link{[<-.dvec}()},
\code{\link{[[.dvec}()},
\code{\link{[[<-.dvec}()},
\code{\link{as.data.frame.dvec}()},
\code{\link{as_dvec}()},
\code{\link{as_dvec.character}()},
\code{\link{as_dvec.dvec}()},
\code{\link{as_dvec.integer}()},
\code{\link{as_dvec.logical}()},
\code{\link{as_dvec.numeric}()},
\code{\link{c.dvec}()},
\code{\link{classified.dvec}()},
\code{\link{desolve.dvec}()},
\code{\link{explicit_guide.data.frame}()},
\code{\link{format.dvec}()},
\code{\link{implicit_guide.dvec}()},
\code{\link{length.dvec}()},
\code{\link{print.dvec}()},
\code{\link{resolve.classified}()},
\code{\link{resolve.dvec}()},
\code{\link{unclassified.dvec}()}
}
\concept{dvec}
\keyword{internal}
