% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{as_dvec.dvec}
\alias{as_dvec.dvec}
\title{Coerce Decorated Vector to Decorated Vector}
\usage{
\method{as_dvec}{dvec}(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{attributes to assign}
}
\value{
dvec
}
\description{
Coerces decorated vector to decorated vector.
Assigns any named attributes in dots.
}
\examples{
as_dvec(as_dvec(letters[1:3]), label = 'Letters')
as_dvec(as_dvec(letters[1:3], label = 'Letters'))
}
\seealso{
Other dvec: 
\code{\link{[.dvec}()},
\code{\link{[<-.dvec}()},
\code{\link{[[.dvec}()},
\code{\link{[[<-.dvec}()},
\code{\link{as.data.frame.dvec}()},
\code{\link{as_dvec.character}()},
\code{\link{as_dvec.complex}()},
\code{\link{as_dvec.integer}()},
\code{\link{as_dvec.logical}()},
\code{\link{as_dvec.numeric}()},
\code{\link{as_dvec}()},
\code{\link{c.dvec}()},
\code{\link{classified.dvec}()},
\code{\link{desolve.dvec}()},
\code{\link{explicit_guide.data.frame}()},
\code{\link{format.dvec}()},
\code{\link{implicit_guide.dvec}()},
\code{\link{length.dvec}()},
\code{\link{print.dvec}()},
\code{\link{resolve.classified}()},
\code{\link{resolve.dvec}()},
\code{\link{unclassified.dvec}()}
}
\concept{dvec}
\keyword{internal}
