% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{as_dvec.integer}
\alias{as_dvec.integer}
\title{Coerce Integer to Decorated Vector}
\usage{
\method{as_dvec}{integer}(x, ...)
}
\arguments{
\item{x}{integer}

\item{...}{ignored arguments}
}
\value{
dvec
}
\description{
Coerces integer to decorated Vector.  Assigns class 'dvec'.
}
\examples{
as_dvec(1:3)
}
\seealso{
Other dvec: 
\code{\link{[.dvec}()},
\code{\link{[<-.dvec}()},
\code{\link{[[.dvec}()},
\code{\link{[[<-.dvec}()},
\code{\link{as_dvec.character}()},
\code{\link{as_dvec.logical}()},
\code{\link{as_dvec.numeric}()},
\code{\link{as_dvec}()},
\code{\link{c.dvec}()},
\code{\link{format.dvec}()},
\code{\link{length.dvec}()},
\code{\link{print.dvec}()}
}
\concept{dvec}
