% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{as_dvec.logical}
\alias{as_dvec.logical}
\title{Coerce Logical to Decorated Vector}
\usage{
\method{as_dvec}{logical}(x, ...)
}
\arguments{
\item{x}{logical}

\item{...}{attributes to assign}
}
\value{
dvec
}
\description{
Coerces logical to decorated vector.
Assigns class 'dvec' and any named attributes in dots.
}
\examples{
as_dvec(c(FALSE, TRUE))
}
\seealso{
Other dvec: 
\code{\link{[.dvec}()},
\code{\link{[<-.dvec}()},
\code{\link{[[.dvec}()},
\code{\link{[[<-.dvec}()},
\code{\link{as.data.frame.dvec}()},
\code{\link{as_dvec.character}()},
\code{\link{as_dvec.complex}()},
\code{\link{as_dvec.dvec}()},
\code{\link{as_dvec.integer}()},
\code{\link{as_dvec.numeric}()},
\code{\link{as_dvec}()},
\code{\link{c.dvec}()},
\code{\link{format.dvec}()},
\code{\link{length.dvec}()},
\code{\link{print.dvec}()}
}
\concept{dvec}
