% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{as_dvec.numeric}
\alias{as_dvec.numeric}
\title{Coerce Numeric to Decorated Vector}
\usage{
\method{as_dvec}{numeric}(x, ...)
}
\arguments{
\item{x}{numeric}

\item{...}{ignored arguments}
}
\value{
dvec
}
\description{
Coerces numeric to decorated Vector.  Assigns class 'dvec'.
}
\examples{
as_dvec(c(10.3, 1.2))
}
\seealso{
Other dvec: 
\code{\link{[.dvec}()},
\code{\link{[<-.dvec}()},
\code{\link{[[.dvec}()},
\code{\link{[[<-.dvec}()},
\code{\link{as_dvec.character}()},
\code{\link{as_dvec.integer}()},
\code{\link{as_dvec.logical}()},
\code{\link{as_dvec}()},
\code{\link{c.dvec}()},
\code{\link{format.dvec}()},
\code{\link{length.dvec}()},
\code{\link{print.dvec}()}
}
\concept{dvec}
