% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{as_dvec}
\alias{as_dvec}
\title{Coerce to Decorated Vector}
\usage{
as_dvec(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{ignored arguments}
}
\value{
dvec
}
\description{
Coerces to Decorated Vector. Generic with methods
\code{\link{as_dvec.integer}},
\code{\link{as_dvec.numeric}},
\code{\link{as_dvec.character}}, and
\code{\link{as_dvec.logical}}.
}
\examples{
as_dvec(0)
}
\seealso{
Other dvec: 
\code{\link{[.dvec}()},
\code{\link{[<-.dvec}()},
\code{\link{[[.dvec}()},
\code{\link{[[<-.dvec}()},
\code{\link{as_dvec.character}()},
\code{\link{as_dvec.integer}()},
\code{\link{as_dvec.logical}()},
\code{\link{as_dvec.numeric}()},
\code{\link{c.dvec}()},
\code{\link{format.dvec}()},
\code{\link{length.dvec}()},
\code{\link{print.dvec}()}
}
\concept{dvec}
