% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{as_expression}
\alias{as_expression}
\title{Coerce to Expression}
\usage{
as_expression(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
expression
}
\description{
Coerces to expression.
Alias for \code{\link{as.expression}}.
}
\keyword{internal}
