% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agplot.R
\name{as_lab.character}
\alias{as_lab.character}
\title{Coerce Character to Axis Label}
\usage{
\method{as_lab}{character}(
  x,
  units = character(0),
  enclose = getOption("yamlet_enclose", default = c("(", ")")),
  parse = getOption("yamlet_label_parse", FALSE),
  ...
)
}
\arguments{
\item{x}{character}

\item{units}{value to use for units}

\item{enclose}{length-two character for enclosing unit}

\item{parse}{whether to convert to expression, or function to do the same}

\item{...}{passed arguments}
}
\value{
character, or expression if parse is TRUE.
}
\description{
Converts character to axis label, optionally returning an expression
intended as \code{\link{plotmath}}.  Respecting units, the intent
is to support udunits syntax (\code{\link[units]{as_units}}).
}
\details{
If \code{parse} is FALSE, \code{units} (if any) are enclosed
in \code{enclose}, and appended to \code{x} with a space
separator. If \code{parse} is a function (or its name), it is passed
all arguments and the result is returned.
if\code{parse} is TRUE, arguments are processed as follows:

* label is converted using \code{\link{as_wikisymbol}} and \code{\link{as_plotmath}},
* spaces in the label are then replaced with plotmath space,
* units if present are converted similarly,
* units will be enclosed using \code{enclose} values with \code{group()},
* label and units(if any) will be combined using \code{paste}, and
* the result is returned as an expression.
}
\seealso{
Other lab: 
\code{\link{as_lab.list}()},
\code{\link{as_lab}()},
\code{\link{singularity}()}
}
\concept{lab}
\keyword{internal}
