% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{as_plotmath.wikisymbol}
\alias{as_plotmath.wikisymbol}
\title{Convert Wiki Symbol to Plotmath}
\usage{
\method{as_plotmath}{wikisymbol}(x, ...)
}
\arguments{
\item{x}{wikisymbol}

\item{...}{ignored}
}
\value{
plotmath
}
\description{
Converts wiki symbol to plotmath. See '?plotmath'.
Vectorized version of \code{\link{wiki_to_plotmath}}.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_wikisymbol(x)
as_plotmath(x)
as_plotmath(as_wikisymbol('gravitational force (kg\\\\.m/s^2.)'))
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol.symbolic_units}()},
\code{\link{as_wikisymbol}()},
\code{\link{wiki_to_latex}()},
\code{\link{wiki_to_plotmath}()}
}
\concept{wikisymbol}
