% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{as_preview.latex}
\alias{as_preview.latex}
\title{Preview Wiki Symbol as Latex}
\usage{
\method{as_preview}{latex}(
  x,
  wide = 50,
  long = 20,
  stem = "latex_preview",
  dir = tempdir(),
  gs_cmd = getOption("gs_cmd", "mgs"),
  ...
)
}
\arguments{
\item{x}{wikisymbol; see \code{\link{as_wikisymbol}}}

\item{wide}{nominal page width}

\item{long}{nominal page length}

\item{dir}{a working directory; see \code{\link[latexpdf]{as.pdf}}}

\item{gs_cmd}{ghostscript command; see \code{\link[latexpdf]{ghostconvert}}}

\item{...}{passed arguments}
}
\value{
invisible filepath
}
\description{
Preview wikisymbol after conversion to latex.
Creates and displays a temporary png file, after
conversion from pdf using \code{\link[latexpdf]{ghostconvert}}.
}
\examples{
\dontrun{
library(magrittr)
'Omega ~ joule^\\\\*. ~1 kg*m^2./s^2' \%>\%
as_wikisymbol \%>\%
as_latex \%>\%
as_preview
}
}
\seealso{
Other preview: 
\code{\link{as_preview.plotmath}()},
\code{\link{as_previews.wikisymbol}()},
\code{\link{as_previews}()},
\code{\link{as_preview}()},
\code{\link{ggplot.plotmath}()},
\code{\link{ggplot.wikisymbol}()}
}
\concept{preview}
\keyword{internal}
