% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-to-plotmath.R
\name{as_preview.plotmath}
\alias{as_preview.plotmath}
\title{Preview Wiki Symbol as Plotmath}
\usage{
\method{as_preview}{plotmath}(x, stem = "plotmath_preview", ...)
}
\arguments{
\item{x}{wikisymbol; see \code{\link{as_wikisymbol}}}

\item{...}{passed arguments}
}
\value{
invisible filepath
}
\description{
Preview wikisymbol after conversion to plotmath.
Creates and displays a temporary png file with
a parsed expression.
}
\examples{
library(magrittr)
'Omega ~ joule^\\\\*. ~1 kg*m^2./s^2' \%>\%
as_wikisymbol \%>\%
as_plotmath \%>\%
as_preview
}
\seealso{
Other preview: 
\code{\link{as_preview.latex}()},
\code{\link{as_preview}()},
\code{\link{ggplot.plotmath}()},
\code{\link{ggplot.wikisymbol}()}
}
\concept{preview}
\keyword{internal}
