% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{as_previews.wikisymbol}
\alias{as_previews.wikisymbol}
\title{Compare Previews of Wiki Symbol}
\usage{
\method{as_previews}{wikisymbol}(x, wide = 13, long = 2.54, width = 0.25, height = 0.05, ...)
}
\arguments{
\item{x}{length-one wikisymbol}

\item{wide}{width in mm of the latex image}

\item{long}{length in mm of the latex image}

\item{width}{width (default: inches) of the plotmath image}

\item{height}{height (default: inches) of the plotmath image}

\item{...}{passed arguments}
}
\value{
invisible list of filepaths
}
\description{
Compares plotmath and latex previews of wikisymbol
Generates png for both, and overlays
latex above plotmath.
}
\examples{
library(magrittr)
specials <- '& \% $ # \\\\_ { } ~ \\\\^ \\\\'
specials \%>\% gsub(' ','',.) \%>\% as_wikisymbol \%>\% as_previews
specials \%>\% as_wikisymbol \%>\% as_previews
as_previews(as_wikisymbol('$'))-> foo
}
\seealso{
Other preview: 
\code{\link{as_preview.latex}()},
\code{\link{as_preview.plotmath}()},
\code{\link{as_previews}()},
\code{\link{as_preview}()},
\code{\link{ggplot.plotmath}()},
\code{\link{ggplot.wikisymbol}()}
}
\concept{preview}
