% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{as_spork.units}
\alias{as_spork.units}
\title{Coerce Units to Spork}
\usage{
\method{as_spork}{units}(x, canonical = TRUE, ...)
}
\arguments{
\item{x}{units; see \code{\link[units]{as_units}}}

\item{...}{ignored arguments}
}
\value{
spork
}
\description{
Coerces units to spork by coercing first
to unit_string.
}
\examples{
library(units)
library(magrittr)
library(spork)
'kg.m^2/s^2' \%>\% as_units \%>\% as_spork
'kg.m2 s-2' \%>\% as_units \%>\% as_spork
'kg.m^2/s^2' \%>\% as_units \%>\% as_spork(FALSE)
'kg.m2 s-2' \%>\% as_units \%>\% as_spork(FALSE)
}
\seealso{
Other spork: 
\code{\link{as_spork.symbolic_units}()},
\code{\link{as_spork.unit_string}()}
}
\concept{spork}
\keyword{internal}
