% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_string.R
\name{as_unit_string.character}
\alias{as_unit_string.character}
\title{Coerce Character to Unit String}
\usage{
\method{as_unit_string}{character}(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored arguments}
}
\value{
unit_string
}
\description{
Coerces character to class 'unit_string'.
See description for \code{\link{as_unit_string}}.
}
\examples{
as_unit_string('kg m2 s-2')
}
\seealso{
Other unit_string: 
\code{\link{[.unit_string}()},
\code{\link{[[.unit_string}()},
\code{\link{as_unit_string.factor}()},
\code{\link{as_unit_string.symbolic_units}()},
\code{\link{as_unit_string.units}()},
\code{\link{as_unit_string}()}
}
\concept{unit_string}
\keyword{internal}
