% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_string.R
\name{as_unit_string}
\alias{as_unit_string}
\title{Coerce to Unit String}
\usage{
as_unit_string(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
unit_string
}
\description{
Coerces to class 'unit_string'. Generic,
with method \code{\link{as_unit_string.character}}.
A unit string is character text suitable
as input for \code{\link[units]{as_units}}.
See also \code{\link{is_parseable}}.
}
\examples{
as_unit_string('kg.m^2/s^2')
as_unit_string('kg.m2 s-2')
}
\seealso{
Other unit_string: 
\code{\link{[.unit_string}()},
\code{\link{[[.unit_string}()},
\code{\link{as_unit_string.character}()},
\code{\link{as_unit_string.factor}()},
\code{\link{as_unit_string.symbolic_units}()},
\code{\link{as_unit_string.units}()}
}
\concept{unit_string}
\keyword{internal}
