% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_string.R
\name{as_unit_string.units}
\alias{as_unit_string.units}
\title{Coerce Units to Unit String.}
\usage{
\method{as_unit_string}{units}(x, canonical = TRUE, ...)
}
\arguments{
\item{x}{units; see \code{\link[units]{as_units}}}

\item{...}{ignored arguments}
}
\value{
unit_string
}
\description{
Coerces units to unit_string. Extracts units
attribute (of class(symbolic_units)) and converts.
}
\examples{
library(units)
x <- as_units('kg.m/s^2')
as_unit_string(x)
as_unit_string(x, canonical = FALSE)

}
\seealso{
Other unit_string: 
\code{\link{[.unit_string}()},
\code{\link{[[.unit_string}()},
\code{\link{as_unit_string.character}()},
\code{\link{as_unit_string.factor}()},
\code{\link{as_unit_string.symbolic_units}()},
\code{\link{as_unit_string}()}
}
\concept{unit_string}
\keyword{internal}
