% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{as_wikisymbol.character}
\alias{as_wikisymbol.character}
\title{Coerce Character to Wiki Symbol}
\usage{
\method{as_wikisymbol}{character}(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored arguments}
}
\value{
wikisymbol
}
\description{
Coerces character to class 'wikisymbol'.
See description for \code{\link{as_wikisymbol}}.
}
\examples{
as_wikisymbol('V_c./F')
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol.symbolic_units}()},
\code{\link{as_wikisymbol}()},
\code{\link{wiki_to_latex}()},
\code{\link{wiki_to_plotmath}()}
}
\concept{wikisymbol}
