% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{as_wikisymbol.factor}
\alias{as_wikisymbol.factor}
\title{Coerce Factor to Wiki Symbol}
\usage{
\method{as_wikisymbol}{factor}(x, ...)
}
\arguments{
\item{x}{factor}

\item{...}{ignored arguments}
}
\value{
wikisymbol
}
\description{
Coerces factor to class 'wikisymbol'
by converting to character and calling
\code{\link{as_wikisymbol}}.
}
\examples{
as_wikisymbol(as.factor('V_c./F'))
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.symbolic_units}()},
\code{\link{as_wikisymbol}()},
\code{\link{wikisym2latex_}()},
\code{\link{wikisym2plotmath_}()}
}
\concept{wikisymbol}
\keyword{internal}
