% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{as_wikisymbol}
\alias{as_wikisymbol}
\title{Coerce to Wiki Symbol}
\usage{
as_wikisymbol(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
wikisymbol
}
\description{
Coerces to class 'wikisymbol'. Generic,
with method \code{\link{as_wikisymbol.character}}.
A Wiki symbol is simple text with arbitrarily
nested subscript (\code{_}) and superscript
(\code{^}) groupings.
}
\details{
Wiki symbol is intended as a compact syntax
that abstracts across differences in
plotmath and latex.  The main purpose
is to handle nested subscripts and superscripts,
but bare words and special symbols are not
disallowed. Support is defined by
\code{\link{as_plotmath}} and
\code{\link{as_latex}}.
Additional rules:
\itemize{
\item Use dot (\code{.}) to explicitly terminate a grouping.
\item Use asterisk(\code{*}) to suggest multiplication.
\item Escape specials with a backslash.
\item Trailing dots need not be supplied.
\item Leading/trailing whitespace is ignored.
\item Internal white space should be respected.
\item Tab character is not allowed.
}
}
\examples{
library(magrittr)
library(dplyr)
library(ggplot2)
library(latexpdf)

data.frame(y=1:10, x=1:10) \%>\%
decorate("x: Omega joule^\\\\*. ~1 kg*m^2./s^2. [\%]") \%>\%
mutate(x = structure(x, label = x \%>\% attr('label') \%>\%
as_wikisymbol \%>\%
as_plotmath \%>\%
as.expression)) \%>\%
ggplot(aes(x, y))

data.frame(y=1:10, x=1:10) \%>\%
decorate("x: gravitational force gamma (kg\\\\.m/s^2.)") \%>\%
mutate(x = structure(x, label = x \%>\% attr('label') \%>\%
as_wikisymbol \%>\%
as_plotmath \%>\%
as.expression)) \%>\%
ggplot(aes(x, y))

path <- tempfile()
data.frame(
 stringsAsFactors = FALSE,
 wikisymbol = c(
   'Omega joule^\\\\*. ~1 kg*m^2./s^2[\%]',
   'gravitational force gamma (kg\\\\.m/s^2.)'
 )
) \%>\%
 transmute(
   latex = wikisymbol \%>\%
   as_wikisymbol \%>\%
   as_latex
 ) \%>\%
 as.pdf(reserve = FALSE, wider = -70)
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol.symbolic_units}()},
\code{\link{wikisym2latex_}()},
\code{\link{wikisym2plotmath_}()}
}
\concept{wikisymbol}
\keyword{internal}
