% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{as_wikisymbol.symbolic_units}
\alias{as_wikisymbol.symbolic_units}
\title{Coerce Symbolic Units to Wiki Symbol.}
\usage{
\method{as_wikisymbol}{symbolic_units}(x, canonical = TRUE, ...)
}
\arguments{
\item{x}{symbolic_units; see \code{\link[units]{as_units}}}

\item{...}{ignored arguments}
}
\value{
wikisymbol
}
\description{
Coerces symbolic units to wikisymbol by coercing first
to unit_string.
}
\examples{
library(units)
x <- as_units('kg.m/s^2')
names(attributes(x))
y <- attr(x,'units')
class(y)
as.character(y)
as.character(attr(x, 'units'))
as_wikisymbol(y)
library(magrittr)
'kg.m^2/s^2' \%>\% as_units \%>\% attr('units') \%>\% as_wikisymbol
'kg.m2 s-2' \%>\% as_units \%>\% attr('units') \%>\% as_wikisymbol
'kg.m^2/s^2' \%>\% as_units \%>\% attr('units') \%>\% as_wikisymbol(FALSE)
'kg.m2 s-2' \%>\% as_units \%>\% attr('units') \%>\% as_wikisymbol(FALSE)
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol.unit_string}()},
\code{\link{as_wikisymbol.units}()},
\code{\link{as_wikisymbol}()},
\code{\link{wiki_to_latex}()},
\code{\link{wiki_to_plotmath}()},
\code{\link{wikitoken}()}
}
\concept{wikisymbol}
\keyword{internal}
