% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{as_wikisymbol.unit_string}
\alias{as_wikisymbol.unit_string}
\title{Coerce Unit String to Wiki Symbol.}
\usage{
\method{as_wikisymbol}{unit_string}(x, ...)
}
\arguments{
\item{x}{unit_string}

\item{...}{ignored arguments}
}
\value{
units
}
\description{
Coerces unit string to wikisymbol.  A literal dot
means different things in wikisymbol vs. units,
and there may be some other subtleties as well.
Unit string is character that \code{\link{is_parseable}}.
}
\examples{
library(magrittr)
'kg.m^2/s^2' \%>\% as_unit_string \%>\% as_wikisymbol
'kg.m2 s-2' \%>\% as_unit_string \%>\% as_wikisymbol
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol.symbolic_units}()},
\code{\link{as_wikisymbol.units}()},
\code{\link{as_wikisymbol}()},
\code{\link{wiki_to_latex}()},
\code{\link{wiki_to_plotmath}()}
}
\concept{wikisymbol}
\keyword{internal}
