% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{as_wikisymbol.units}
\alias{as_wikisymbol.units}
\title{Coerce Units to Wiki Symbol.}
\usage{
\method{as_wikisymbol}{units}(x, canonical = TRUE, ...)
}
\arguments{
\item{x}{units; see \code{\link[units]{as_units}}}

\item{...}{ignored arguments}
}
\value{
wikisymbol
}
\description{
Coerces units to wikisymbol by coercing first
to unit_string.
}
\examples{
library(units)
library(magrittr)
'kg.m^2/s^2' \%>\% as_units \%>\% as_wikisymbol
'kg.m2 s-2' \%>\% as_units \%>\% as_wikisymbol
'kg.m^2/s^2' \%>\% as_units \%>\% as_wikisymbol(FALSE)
'kg.m2 s-2' \%>\% as_units \%>\% as_wikisymbol(FALSE)
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol.symbolic_units}()},
\code{\link{as_wikisymbol.unit_string}()},
\code{\link{as_wikisymbol}()},
\code{\link{wiki_to_latex}()},
\code{\link{wiki_to_plotmath}()},
\code{\link{wikitoken}()}
}
\concept{wikisymbol}
\keyword{internal}
