% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{as_yam.character}
\alias{as_yam.character}
\title{Coerce Character to yam}
\usage{
\method{as_yam}{character}(x, as.named.list, ...)
}
\arguments{
\item{x}{length-one filepath or actual data}

\item{as.named.list}{enforced as TRUE}

\item{...}{passed to \code{\link[yaml]{read_yaml}}}
}
\value{
a named list
}
\description{
Coerces character to yam.  Length-one character can be
a filepath, otherwise treated as data.  Proceeds by
importing the data and determining the default keys.
}
\examples{
as_yam(system.file(package = 'yamlet', 'extdata','yamlet.yaml'))
as_yam(c('ID:','TIME:'))
as_yam('ID:\nTIME:')

}
\seealso{
Other yam: 
\code{\link{as.character.yam}()},
\code{\link{as_yam.yamlet}()},
\code{\link{as_yam}()}
}
\concept{yam}
