% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{as_yam}
\alias{as_yam}
\title{Convert Stored YAML to Intermediate Form}
\usage{
as_yam(file, as.named.list, ...)
}
\arguments{
\item{file}{passed to \code{\link[yaml]{read_yaml}}}

\item{as.named.list}{enforced as TRUE}

\item{...}{passed to \code{\link[yaml]{read_yaml}}}
}
\value{
a named list
}
\description{
Converts stored YAML to intermediate form by
importing the data and determining the default keys.
}
\examples{
as_yam(system.file(package = 'yamlet', 'extdata','yamlet.yaml'))

}
\seealso{
Other yamlet: \code{\link{as_yamlet.character}},
  \code{\link{as_yamlet.yam}}, \code{\link{as_yamlet}}
}
\concept{yamlet}
