% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{as_yamlet.data.frame}
\alias{as_yamlet.data.frame}
\title{Coerce Data.frame to Yamlet}
\usage{
\method{as_yamlet}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{passed arguments}
}
\value{
yamlet
}
\description{
Coerces data.frame to yamlet. Assigns class 'yamlet' do a data.frame's decorations.
}
\examples{
library(csv)
file <- system.file(package = 'yamlet', 'extdata','yamlet.csv')
x <- decorate(as.csv(file))
as_yamlet(x)
}
\concept{as_yamlet}
