% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{as_yamlet}
\alias{as_yamlet}
\title{Convert To yamlet Format}
\usage{
as_yamlet(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Converts something to yamlet format. If the object
or user specifies default keys, these are applied,
with the former having priority.  See \code{\link{as_yamlet.character}}.
}
\examples{
as_yamlet(as_yam(system.file(package = 'yamlet', 'extdata','yamlet.yaml')))

}
\seealso{
Other yamlet: \code{\link{as_yamlet.character}},
  \code{\link{as_yamlet.yam}}, \code{\link{as_yam}}
}
\concept{yamlet}
