% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{as_yamlet.yam}
\alias{as_yamlet.yam}
\title{Coerce Yam To Yamlet Format}
\usage{
\method{as_yamlet}{yam}(x, default_keys = list("label", "guide"), ...)
}
\arguments{
\item{x}{a yam object; see \code{\link{as_yam}}}

\item{default_keys}{character: default keys for the first n anonymous members of each element}

\item{...}{passed arguments}
}
\value{
yamlet: a named list with default keys applied
}
\description{
Coerces yam to yamlet format. If the object
or user specifies default keys, these are applied,
with the former having priority.  See \code{\link{as_yamlet.character}}.
}
\examples{
as_yamlet(as_yam(system.file(package = 'yamlet', 'extdata','yamlet.yaml')))

}
\seealso{
Other yamlet: \code{\link{as_yamlet.character}},
  \code{\link{as_yamlet}}
}
\concept{yamlet}
