% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{as_yamlet.yamlet}
\alias{as_yamlet.yamlet}
\title{Coerce Yamlet to Yamlet}
\usage{
\method{as_yamlet}{yamlet}(x, ...)
}
\arguments{
\item{x}{yamlet}

\item{...}{ignored}
}
\value{
yamlet
}
\description{
Coerces yamlet to yamlet. Currently a non-operation.
}
\examples{
meta <- system.file(package = 'yamlet', 'extdata','quinidine.yaml')
x <- as_yamlet(meta)
as_yamlet(x)
}
\seealso{
Other yamlet: 
\code{\link{[.yamlet}()},
\code{\link{as.character.yamlet}()},
\code{\link{as_yamlet.character}()},
\code{\link{as_yamlet.data.frame}()},
\code{\link{as_yamlet.list}()},
\code{\link{as_yamlet.yam}()},
\code{\link{as_yamlet}()},
\code{\link{print.yamlet}()}
}
\concept{yamlet}
\keyword{internal}
