% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{canonical}
\alias{canonical}
\title{Enforce Canonical Order}
\usage{
canonical(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
list
}
\description{
Enforce canonical order.  Generic, with
method: \code{\link{canonical.decorated}}.
}
\seealso{
Other canonical: 
\code{\link{canonical.decorated}()},
\code{\link{canonical.yamlet}()}
}
\concept{canonical}
\keyword{internal}
