% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{cell_value}
\alias{cell_value}
\title{Calculate a Cell Value}
\usage{
cell_value(x, sep = "\\n", def = ": ")
}
\arguments{
\item{x}{list of character, possibly named}

\item{sep}{separator for multiple items within an attribute}

\item{def}{definition string: separator between items and their (preceding) names, if any}
}
\description{
Calculates a cell value.
}
\keyword{internal}
