% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified.R
\name{classified.classified}
\alias{classified.classified}
\title{Create Classified from Classified}
\usage{
\method{classified}{classified}(x, ...)
}
\arguments{
\item{x}{classified}

\item{...}{ignored}
}
\value{
'classified' 'factor'
}
\description{
See \code{\link{classified.default}}.
Currently, calling classified on a
classified object is a non-operation.
}
\examples{

a <- 4:6
attr(a, 'codelist') <- list(d = 4, e = 5, f = 6, g = 7)
b <- classified(a)
a
b
class(b)
classified(b)
identical(b, classified(b))
}
\seealso{
Other classified: 
\code{\link{[.classified}()},
\code{\link{[<-.classified}()},
\code{\link{[[.classified}()},
\code{\link{[[<-.classified}()},
\code{\link{as.integer.classified}()},
\code{\link{c.classified}()},
\code{\link{classified.data.frame}()},
\code{\link{classified.default}()},
\code{\link{classified.dvec}()},
\code{\link{classified}()},
\code{\link{desolve.classified}()},
\code{\link{unclassified.classified}()},
\code{\link{unclassified.data.frame}()},
\code{\link{unclassified}()}
}
\concept{classified}
\keyword{internal}
