% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified.R
\name{classified.data.frame}
\alias{classified.data.frame}
\title{Classify Data Frame}
\usage{
\method{classified}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{passed to \code{\link[dplyr]{select}} to limit scope}
}
\value{
data.frame
}
\description{
Coerces items in data.frame with codelist attribute to 'classified':
a factor with a codelist attribute.
}
\examples{
library(magrittr)
file <- system.file(package = 'yamlet', 'extdata','quinidine.csv')
x <- decorate(file)
x \%>\% explicit_guide \%>\% decorations(Age, Race, Heart:glyco)
x \%>\% explicit_guide \%>\% classified \%>\% decorations(Age, Race, Heart:glyco)
x \%>\% explicit_guide \%>\% classified(Heart:glyco) \%>\% decorations(Age, Race, Heart:glyco)
}
\seealso{
Other classified: 
\code{\link{[.classified}()},
\code{\link{[<-.classified}()},
\code{\link{[[.classified}()},
\code{\link{[[<-.classified}()},
\code{\link{c.classified}()},
\code{\link{classified.default}()},
\code{\link{classified}()},
\code{\link{unclassified.classified}()},
\code{\link{unclassified.data.frame}()},
\code{\link{unclassified}()}
}
\concept{classified}
\keyword{internal}
