% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified.R
\name{classified}
\alias{classified}
\title{Classify Something.}
\usage{
classified(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Classifies something.
Generic, with method \code{\link{classified.default}}
}
\examples{
example(classified.default)
}
\keyword{internal}
