% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelist.R
\name{codelist.default}
\alias{codelist.default}
\title{Create a Codelist by Default}
\usage{
\method{codelist}{default}(x, levels, ...)
}
\arguments{
\item{x}{vector or factor}

\item{...}{passed to \code{\link{factor}}}

\item{names}{optional}
}
\value{
list, possibly named
}
\description{
Creates a codelist by default.
Coerces to factor, and supplies
levels as names of corresponding
elements.
}
\examples{
library(magrittr)

# Suppose we have a numeric and a character version of treatment:

x <- data.frame(
  TRTN =  c(        0,      1,            2,         0),
  TRTC =  c('placebo', 'drug', 'comparitor', 'placebo')
)

x

# Further, we wish to specify a particular display order:

x$TRTC <- factor(x$TRTC, levels = c('drug', 'comparitor', 'placebo'))

# We can use TRTC as the guide for TRTN:

x \%<>\% modify(TRTN, guide = codelist(x, TRTC, TRTN))

x \%>\% decorations(TRTN)


}
\seealso{
Other codelist: 
\code{\link{codelist.data.frame}()},
\code{\link{codelist}()}
}
\concept{codelist}
\keyword{internal}
