% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelist.R
\name{codelist}
\alias{codelist}
\title{Create a Codelist}
\usage{
codelist(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Creates a codelist.  Generic, with method \code{\link{codelist.data.frame}}
}
\examples{
example(codelist.data.frame)
}
\seealso{
Other codelist: 
\code{\link{codelist.data.frame}()},
\code{\link{codelist.default}()}
}
\concept{codelist}
\keyword{internal}
