% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{collapse.default}
\alias{collapse.default}
\title{Collapse Uninformative Levels by Default}
\usage{
\method{collapse}{default}(x)
}
\arguments{
\item{x}{object}
}
\description{
The default collapse() method returns the unmodified object.
}
\examples{
collapse(yaml::yaml.load('ITEM:'))
}
\seealso{
Other collapse: \code{\link{collapse.list}},
  \code{\link{collapse}}
}
\concept{collapse}
\keyword{internal}
