% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{collapse.list}
\alias{collapse.list}
\title{Collapse Uninformative Levels of a List}
\usage{
\method{collapse}{list}(x)
}
\arguments{
\item{x}{list}
}
\value{
list
}
\description{
The list method for collapse() recursively
ascends a nested list, removing uninformative levels.
}
\examples{
collapse(yaml::yaml.load('ITEM: [ label: sunshine, [foo: 1, bar: 3]]'))
}
\seealso{
Other collapse: \code{\link{collapse.default}},
  \code{\link{collapse}}
}
\concept{collapse}
\keyword{internal}
