% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{collapse}
\alias{collapse}
\title{Collapse Uninformative Levels}
\usage{
collapse(x)
}
\arguments{
\item{x}{object}
}
\description{
Collapses uninformative levels of nested lists.
If any element of a list (recursively) is a list
with one member, that member replaces the element,
combining the element's name (if any) with the
member's name (if any).
}
\examples{
collapse(yaml::yaml.load('[foo: 1, bar: 3]'))
}
\seealso{
Other collapse: \code{\link{collapse.default}},
  \code{\link{collapse.list}}
}
\concept{collapse}
\keyword{internal}
