% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalize.R
\name{conditionalize}
\alias{conditionalize}
\title{Conditionalize Attributes}
\usage{
conditionalize(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Conditionalizes attributes of something.
Generic, with method for data.frame.
}
\examples{
example(conditionalize.data.frame)
}
\seealso{
Other conditionalize: 
\code{\link{conditionalize.data.frame}()}
}
\concept{conditionalize}
\keyword{internal}
