% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapframe.R
\name{data_context.ggplot}
\alias{data_context.ggplot}
\title{Retrieve Data Context of ggplot}
\usage{
\method{data_context}{ggplot}(x, ...)
}
\arguments{
\item{x}{ggplot}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Retrieves data context of ggplot as 'data.frame'.
calls \code{\link{mapframe.ggplot}}, 
and limits this to rows where 'data' is not NA
and 'aesthetic' is not duplicated.
I.e., for each aesthetic, this function 
returns the first mapping that has (at least default)
data.  Such mappings should be what ggplot uses
to train the corresponding scales.
}
\seealso{
Other decorated_ggplot: 
\code{\link{data_context}()},
\code{\link{ggplot.decorated}()},
\code{\link{ggplot_build.decorated_ggplot}()},
\code{\link{mapframe}()},
\code{\link{mapframe.default}()},
\code{\link{mapframe.ggplot}()}
}
\concept{decorated_ggplot}
\keyword{internal}
