% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapframe.R
\name{data_context}
\alias{data_context}
\title{Retrieve Data Context}
\usage{
data_context(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
Retrieves data context as 'data.frame'.
Generic, with methods \code{\link{data_context.ggplot}}.
}
\seealso{
Other decorated_ggplot: 
\code{\link{data_context.ggplot}()},
\code{\link{ggplot.decorated}()},
\code{\link{ggplot_build.decorated_ggplot}()},
\code{\link{mapframe}()},
\code{\link{mapframe.default}()},
\code{\link{mapframe.ggplot}()}
}
\concept{decorated_ggplot}
\keyword{internal}
