% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_decorated.R
\name{dd_cast}
\alias{dd_cast}
\title{Coerce Data.frame to Decorated}
\usage{
dd_cast(x, to, ...)
}
\arguments{
\item{x}{subclass of data.frame}

\item{to}{subclass of data.frame}

\item{...}{passed arguments}
}
\value{
decorated
}
\description{
Coerces data.frame to decorated.  Wrapper for \code{\link[vctrs]{df_cast}}.
}
