% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_decorated.R
\name{dd_ptype2}
\alias{dd_ptype2}
\title{Coerce Common Type to Decorated}
\usage{
dd_ptype2(x, y, ...)
}
\arguments{
\item{x}{subclass of data.frame}

\item{y}{subclass of data.frame}

\item{...}{passed arguments}
}
\value{
decorated
}
\description{
Coerces common type to decorated.  Wrapper for \code{\link[vctrs]{df_ptype2}}.
}
