% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{decorate}
\alias{decorate}
\title{Decorate a List-like Object}
\usage{
decorate(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
a list-like object, typically data.frame
}
\description{
Decorates a list-like object. Generic.  See \code{\link{decorate.character}}.
}
\examples{
library(csv)
file <- system.file(package = 'yamlet', 'extdata','yamlet.csv')
x <- decorate(as.csv(file))
sapply(x, attr, 'label')

}
\seealso{
Other decorate: \code{\link{decorate.character}},
  \code{\link{decorate.data.frame}},
  \code{\link{decorate.list}}
}
\concept{decorate}
