% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{deflate.default}
\alias{deflate.default}
\title{Collapse Uninformative Levels by Default}
\usage{
\method{deflate}{default}(x)
}
\arguments{
\item{x}{object}
}
\value{
list
}
\description{
The default deflate() method returns the unmodified object.
}
\examples{
deflate(yaml::yaml.load('ITEM:'))
}
\seealso{
Other deflate: \code{\link{deflate.list}},
  \code{\link{deflate}}
}
\concept{deflate}
\keyword{internal}
