% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{deflate.list}
\alias{deflate.list}
\title{Collapse Uninformative Levels of a List}
\usage{
\method{deflate}{list}(x)
}
\arguments{
\item{x}{list}
}
\value{
list
}
\description{
The list method for deflate() recursively
ascends a nested list, removing uninformative levels.
}
\examples{
deflate(yaml::yaml.load('ITEM: [ label: sunshine, [foo: 1, bar: 3]]'))
}
\seealso{
Other deflate: \code{\link{deflate.default}},
  \code{\link{deflate}}
}
\concept{deflate}
\keyword{internal}
