% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{deflate}
\alias{deflate}
\title{Collapse Uninformative Levels}
\usage{
deflate(x)
}
\arguments{
\item{x}{object}
}
\value{
list
}
\description{
Collapses uninformative levels of nested lists.
If any element of a list (recursively) is a list
with one member, that member replaces the element,
combining the element's name (if any) with the
member's name (if any).
}
\examples{
deflate(yaml::yaml.load('[foo: 1, bar: 3]'))
}
\seealso{
Other deflate: 
\code{\link{deflate.default}()},
\code{\link{deflate.list}()}
}
\concept{deflate}
\keyword{internal}
