% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desolve.R
\name{desolve.data.frame}
\alias{desolve.data.frame}
\title{Desolve Data Frame}
\usage{
\method{desolve}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
decorated
}
\description{
Desolves data.frame.
Coerces first using as_decorated().
}
\examples{
head(desolve(Theoph))
}
\seealso{
Other resolve: 
\code{\link{desolve.classified}()},
\code{\link{desolve.decorated}()},
\code{\link{desolve.dvec}()},
\code{\link{desolve}()},
\code{\link{resolve.classified}()},
\code{\link{resolve.data.frame}()},
\code{\link{resolve.decorated}()},
\code{\link{resolve.dvec}()},
\code{\link{resolve}()}
}
\concept{resolve}
\keyword{internal}
