% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desolve.R
\name{desolve.default}
\alias{desolve.default}
\title{Desolve Default}
\usage{
\method{desolve}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
class of x
}
\description{
By default, desolves by returning the unmodified object.
}
\seealso{
Other resolve: 
\code{\link{desolve}()},
\code{\link{desolve.classified}()},
\code{\link{desolve.data.frame}()},
\code{\link{desolve.decorated}()},
\code{\link{desolve.dvec}()},
\code{\link{resolve}()},
\code{\link{resolve.classified}()},
\code{\link{resolve.data.frame}()},
\code{\link{resolve.decorated}()},
\code{\link{resolve.default}()},
\code{\link{resolve.dvec}()},
\code{\link{resolve.factor}()}
}
\concept{resolve}
\keyword{internal}
