% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_title.R
\name{drop_title.default}
\alias{drop_title.default}
\title{Drop Title by Default}
\usage{
\method{drop_title}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\value{
same as x
}
\description{
Drops title by default.
To be specific:  this is the default method
for the generic function \code{\link{drop_title}},
and it actually does nothing.  Individual methods
are written for those classes where 'drop title' 
behavior is expected.
}
\seealso{
Other labels: 
\code{\link{alias.data.frame}()},
\code{\link{append_units.data.frame}()},
\code{\link{append_units.default}()},
\code{\link{append_units}()},
\code{\link{drop_title.decorated}()},
\code{\link{drop_title.dvec}()},
\code{\link{drop_title}()},
\code{\link{make_title.decorated}()},
\code{\link{make_title.default}()},
\code{\link{make_title.dvec}()},
\code{\link{make_title}()},
\code{\link{sub_units}()}
}
\concept{labels}
\keyword{internal}
