% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{encode.yamlet}
\alias{encode.yamlet}
\title{Encode yamelet}
\usage{
\method{encode}{yamlet}(x, target = "guide", ...)
}
\arguments{
\item{x}{yamlet}

\item{target}{attribute to encode}

\item{...}{ignored}
}
\value{
yamlet, with guide elements possibly transformed to encodings
}
\description{
Encodes yamlet.  Each 'guide' element that is a list
is converted to an encoding, if possible.
}
\examples{
meta <- system.file(package = 'yamlet', 'extdata','yamlet.yaml')
meta <- as_yamlet(meta)
meta <- encode(meta)
}
\seealso{
Other encode: \code{\link{[.yamlet}},
  \code{\link{list2encoding}}
}
\concept{encode}
