% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enscript.R
\name{enscript}
\alias{enscript}
\title{Render Scripted Attributes of Indicated Components}
\usage{
enscript(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Renders the scripted attributes of indicated components.
Generic, with method \code{\link{enscript.default}}.
}
\examples{
example(enscript.default)
}
\seealso{
modify.default

as_spork

Other enscript: 
\code{\link{enscript.default}()}
}
\concept{enscript}
\keyword{internal}
