% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_label_units.R
\name{express_label_units.default}
\alias{express_label_units.default}
\title{Convert Label to Expression}
\usage{
\method{express_label_units}{default}(x, wrap = c(" (", ")"), ...)
}
\arguments{
\item{x}{object}

\item{wrap}{character to wrap units (will be coerced to length two)}

\item{...}{passed arguments}
}
\value{
like x with label attribute of class 'expression'
}
\description{
Converts Label to Expression.
Joins with units by default, and coerces to
spork, plotmath, and expression.
}
\examples{
library(units)
library(magrittr)
x <- 1:10
attr(x, 'label') <- 'acceleration'
units(x) <- 'm/s^2' # using units package
y <- as_units('kg') # using units package
x \%>\% attr('label')
x \%>\% express_label_units \%>\% attr('label')
y \%>\% attr('label')
y \%>\% express_label_units \%>\% attr('label')
express_label_units(2)
z <- 1:10
attr(z, 'label') <- 'joules'
attr(z, 'units') <- 'kg.m^2./s^2' # vanilla units
express_label_units(z, wrap = c(' [',']'))
a <- 3
attr(a, 'label') <- 'C_max_ss'
attr(a, 'units') <- 'ng/mL_o'
a \%>\% express_label_units \%>\% attr('label')
}
\seealso{
Other labels: 
\code{\link{express_label_units.data.frame}()},
\code{\link{express_label_units}()},
\code{\link{join_label_units.data.frame}()},
\code{\link{join_label_units.default}()},
\code{\link{join_label_units}()}
}
\concept{labels}
