% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_label_units.R
\name{express_label_units}
\alias{express_label_units}
\title{Convert Label and Units to Expression}
\usage{
express_label_units(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Converts Label and Units to Expression.
Generic, with default method.
}
\examples{
example(express_label.default)
}
\seealso{
Other labels: 
\code{\link{express_label_units.data.frame}()},
\code{\link{express_label_units.default}()},
\code{\link{join_label_units.data.frame}()},
\code{\link{join_label_units.default}()},
\code{\link{join_label_units}()}
}
\concept{labels}
\keyword{internal}
