% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_codelist.R
\name{factorize_codelist.character}
\alias{factorize_codelist.character}
\title{Coerce Character with Codelist to Factor}
\usage{
\method{factorize_codelist}{character}(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{passed arguments}
}
\value{
class 'classified' 'factor'
}
\description{
Coerces character with codelist attribute to factor.
If attribute 'codelist' is missing, unique values of
x are supplied.
}
\examples{
example(factorize_codelist.data.fame)
}
\seealso{
Other factorize_codelist: 
\code{\link{factorize_codelist.data.frame}()},
\code{\link{factorize_codelist.default}()},
\code{\link{factorize_codelist.factor}()},
\code{\link{factorize_codelist}()}
}
\concept{factorize_codelist}
\keyword{internal}
