% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_codelist.R
\name{factorize_codelist.factor}
\alias{factorize_codelist.factor}
\title{Coerce Codelist to Factor for Factor}
\usage{
\method{factorize_codelist}{factor}(x, ...)
}
\arguments{
\item{x}{presumably vector-like or factor}

\item{...}{passed arguments}
}
\value{
factor
}
\description{
Coerces Codelist to Factor for Factors.
Coerces to character and calls next method.
}
\seealso{
Other factorize_codelist: 
\code{\link{factorize_codelist.character}()},
\code{\link{factorize_codelist.data.frame}()},
\code{\link{factorize_codelist.default}()},
\code{\link{factorize_codelist}()}
}
\concept{factorize_codelist}
\keyword{internal}
