% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_codelist.R
\name{factorize_codelist}
\alias{factorize_codelist}
\title{Coerce Codelist to Factor}
\usage{
factorize_codelist(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Coerces codelist to factor.  Generic, with default and data.frame methods.
}
\seealso{
Other factorize_codelist: 
\code{\link{factorize_codelist.character}()},
\code{\link{factorize_codelist.data.frame}()},
\code{\link{factorize_codelist.default}()},
\code{\link{factorize_codelist.factor}()}
}
\concept{factorize_codelist}
