% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{filter.decorated}
\alias{filter.decorated}
\title{Filter Decorated}
\usage{
\method{filter}{decorated}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Expressions that return a
logical value, and are defined in terms of the variables in \code{.data}.
If multiple expressions are included, they are combined with the \code{&} operator.
Only rows for which all conditions evaluate to \code{TRUE} are kept.}

\item{.preserve}{Relevant when the \code{.data} input is grouped.
If \code{.preserve = FALSE} (the default), the grouping structure
is recalculated based on the resulting data, otherwise the grouping is kept as is.}
}
\description{
Preserves class when filtering decorated.
}
\seealso{
Other dplyr: 
\code{\link{anti_join.decorated}()},
\code{\link{arrange.decorated}()},
\code{\link{as_decorated}()},
\code{\link{full_join.decorated}()},
\code{\link{group_by.decorated}()},
\code{\link{inner_join.decorated}()},
\code{\link{left_join.decorated}()},
\code{\link{mutate.decorated}()},
\code{\link{right_join.decorated}()},
\code{\link{select.decorated}()},
\code{\link{semi_join.decorated}()},
\code{\link{slice.decorated}()},
\code{\link{summarise.decorated}()},
\code{\link{summarize.decorated}()}
}
\concept{dplyr}
\keyword{internal}
