% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable.R
\name{footnote.decorated}
\alias{footnote.decorated}
\title{Footnote Decorated.}
\usage{
\method{footnote}{decorated}(x, equal = ":", collapse = "; ", ...)
}
\arguments{
\item{x}{decorated}

\item{equal}{character: a symbol suggesting equality between a name and its note}

\item{collapse}{used to \code{\link{paste}} column-wise footnotes}

\item{...}{passed to append_units}
}
\value{
character
}
\description{
Footnotes a decorated data.frame.
Generates a text string that defines
column names using label and unit attributes.
}
\examples{
library(magrittr)
set.seed(0)
x <- data.frame(
 auc = rnorm(100, mean = 2400, sd = 200),
 bmi = rnorm(100, mean = 20, sd = 5),
 gen = 0:1
)
}
\seealso{
Other footnote: 
\code{\link{footnote}()}
}
\concept{footnote}
