% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable.R
\name{footnote}
\alias{footnote}
\title{Footnote Something.}
\usage{
footnote(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Footnotes something.
Generic, with method \code{\link{footnote.decorated}}.
}
\examples{
# see methods
}
\seealso{
Other footnote: 
\code{\link{footnote.decorated}()}
}
\concept{footnote}
\keyword{internal}
