% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{gather.decorated}
\alias{gather.decorated}
\title{Gather a Decorated Data Frame}
\usage{
\method{gather}{decorated}(
  data,
  key = "key",
  value = "value",
  ...,
  na.rm = FALSE,
  convert = FALSE,
  factor_key = FALSE
)
}
\arguments{
\item{data}{see \code{\link[tidyr]{gather}}}

\item{key}{see \code{\link[tidyr]{gather}}}

\item{value}{see \code{\link[tidyr]{gather}}}

\item{...}{see \code{\link[tidyr]{gather}}}

\item{na.rm}{see \code{\link[tidyr]{gather}}}

\item{convert}{see \code{\link[tidyr]{gather}}}

\item{factor_key}{see \code{\link[tidyr]{gather}}}
}
\value{
decorated
}
\description{
Gathers a decorated data.frame.
I.e. a gather method for class 'decorated'.
Invokes tidyr::gather(), converting gathered
labels to the guide attribute of \code{key}, and
converting gathered guides ... if all the same ...
to the guide attribute of \code{value}.
Somewhat experimental!
}
\examples{
library(magrittr)
library(tidyr)
file <- system.file(package = 'yamlet', 'extdata','quinidine.csv')
x <- decorate(file)
x \%>\% gather('key', 'value', time, interval) \%>\% decorations

}
\keyword{internal}
