% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{get_labs.ggplot}
\alias{get_labs.ggplot}
\title{Get Labels for ggplot}
\usage{
\method{get_labs}{ggplot}(plot = get_last_plot())
}
\arguments{
\item{plot}{the ggplot}
}
\description{
Gets labels for a ggplot. Returns plot$labels for 3.5.1 and earlier,
else returns ggplot2::get_labs.
}
