% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{get_labs.ggplot_decorated}
\alias{get_labs.ggplot_decorated}
\title{Get Labels for ggplot_decorated}
\usage{
\method{get_labs}{ggplot_decorated}(plot)
}
\arguments{
\item{plot}{the ggplot}
}
\description{
Gets labels for a ggplot_decorated.  Calls next method to
avoid endless loop (since ggplot_build.ggplot_decorated calls get_labs() indirectly).
}
