% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{get_labs}
\alias{get_labs}
\title{Get Labels}
\usage{
get_labs(plot = get_last_plot())
}
\description{
Gets labels for a ggplot object.  Not exported, to avoid confusion.
Development version of ggplot2 implements new get_labs() and get_strip_labels()
interfaces. This function is an abstraction that supports new vs old approaches.
See https://github.com/tidyverse/ggplot2/pull/6078.
}
