% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{gg_new}
\alias{gg_new}
\title{Detect Revised Label Strategy}
\usage{
gg_new()
}
\description{
Detects the existence of qqplot's updated label strategy after v. 3.5.1,
e.g. ggplot2_3.5.1.9000. For internal use to accommodate breaking changes.
}
