% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot.resolved}
\alias{ggplot.resolved}
\title{Create a New ggplot for a Resolved Data Frame}
\usage{
\method{ggplot}{resolved}(data, ...)
}
\arguments{
\item{data}{data.frame or similar; class will be coerced to data.frame}

\item{...}{passed to \code{\link[ggplot2]{ggplot}}}
}
\value{
return value like \code{\link[ggplot2]{ggplot}} but inheriting 'rg'
}
\description{
Creates a new ggplot object for a decorated data.frame.
This is the ggplot() method for class 'resolved'.
It creates a ggplot object with a data.frame version of \code{x},
but reclassifies it as 'rg' so that a custom print method
is invoked; see \code{\link{print.rg}}.
}
\examples{
example(print.rg)
}
\seealso{
Other dg: 
\code{\link{ggplot.ggready}()},
\code{\link{ggplot_build.dg}()},
\code{\link{ggplot_build.rg}()},
\code{\link{print.dg}()},
\code{\link{print.rg}()}

Other interface: 
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{ggready.data.frame}()},
\code{\link{io_csv.character}()},
\code{\link{io_csv.data.frame}()},
\code{\link{io_res.character}()},
\code{\link{io_table.character}()},
\code{\link{io_table.data.frame}()},
\code{\link{io_yamlet.character}()},
\code{\link{io_yamlet.data.frame}()},
\code{\link{is_parseable.default}()},
\code{\link{modify.default}()},
\code{\link{read_yamlet}()},
\code{\link{resolve.decorated}()},
\code{\link{selected.default}()},
\code{\link{write_yamlet}()}
}
\concept{dg}
\concept{interface}
