% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-to-plotmath.R
\name{ggplot.wikisymbol}
\alias{ggplot.wikisymbol}
\title{Plot Wiki Symbol}
\usage{
\method{ggplot}{wikisymbol}(x, blank = TRUE, ...)
}
\arguments{
\item{x}{length-one wikisymbol; see \code{\link{as_wikisymbol}}}

\item{blank}{whether to use a blank plot area}

\item{...}{ignored arguments}
}
\value{
gg
}
\description{
Render wikisymbol in a ggplot.
}
\examples{
'1 joule^\\\\*. ~1 kg m^2./s^2' \%>\% as_wikisymbol \%>\% ggplot
}
\seealso{
Other preview: 
\code{\link{as_preview.latex}()},
\code{\link{as_preview.plotmath}()},
\code{\link{as_previews.wikisymbol}()},
\code{\link{as_previews}()},
\code{\link{as_preview}()},
\code{\link{ggplot.plotmath}()}
}
\concept{preview}
\keyword{internal}
