% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isometric.R
\name{ggplot_add.ggplot_isometric}
\alias{ggplot_add.ggplot_isometric}
\title{Add Isometry to Plot Object}
\usage{
\method{ggplot_add}{ggplot_isometric}(object, plot, object_name)
}
\value{
gg
}
\description{
Adds isometry to plot object.
}
\examples{
example(isometric)
}
\seealso{
isometric

Other isometric: 
\code{\link{ggplot_add.ggplot_symmetric}()},
\code{\link{isometric}()},
\code{\link{symmetric}()}
}
\concept{isometric}
\keyword{internal}
