% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_striptext.R
\name{ggplot_add.ggplot_scale_striptext}
\alias{ggplot_add.ggplot_scale_striptext}
\title{Adjust Striptext Font Size}
\usage{
\method{ggplot_add}{ggplot_scale_striptext}(object, plot, object_name)
}
\arguments{
\item{object}{length-one non-negative real number controling severity of scale adjustment}

\item{plot}{ggplot}

\item{object_name}{unused}
}
\value{
gg
}
\description{
Adjust striptext font size for class 'ggplot'.
}
\examples{
example(scale_striptext)
}
\seealso{
Other isometric: 
\code{\link{ggplot_add.ggplot_isometric}()},
\code{\link{ggplot_add.ggplot_symmetric}()},
\code{\link{isometric}()},
\code{\link{scale_striptext}()},
\code{\link{symmetric}()}
}
\concept{isometric}
\keyword{internal}
