% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isometric.R
\name{ggplot_add.ggplot_symmetric}
\alias{ggplot_add.ggplot_symmetric}
\title{Add Symmetry to Plot Object}
\usage{
\method{ggplot_add}{ggplot_symmetric}(object, plot, object_name)
}
\value{
gg
}
\description{
Adds y axis symmetry to plot object.
}
\examples{
example(symmetric)
}
\seealso{
symmetric

Other isometric: 
\code{\link{ggplot_add.ggplot_isometric}()},
\code{\link{isometric}()},
\code{\link{symmetric}()}
}
\concept{isometric}
\keyword{internal}
