% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_ggready.R
\name{ggplot_build.ggready_ggplot}
\alias{ggplot_build.ggready_ggplot}
\title{Enable Automatic Labels and Units for ggplot}
\usage{
ggplot_build.ggready_ggplot(x, ...)
}
\arguments{
\item{x}{class 'ggready_ggplot' from \code{\link{ggplot.ggready}}}

\item{...}{passed arguments}
}
\value{
see \code{\link[ggplot2]{ggplot_build}}
}
\description{
Enable automatic labels and units for ggplot.
Substitutes column label, if present, for default.
Supports arrangements of ggplot objects.
}
\seealso{
Other ggready: 
\code{\link{ggplot.ggready}()},
\code{\link{ggready.data.frame}()},
\code{\link{ggready.decorated}()},
\code{\link{ggready.resolved}()},
\code{\link{ggready}()},
\code{\link{print.ggready_ggplot}()}
}
\concept{ggready}
\keyword{internal}
