% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggready.R
\name{ggready.decorated}
\alias{ggready.decorated}
\title{Prepare Decorated Data Frame for GGplot}
\usage{
\method{ggready}{decorated}(x, ..., parse = getOption("ggready_parse", TRUE))
}
\arguments{
\item{x}{object}

\item{...}{passed to \code{\link{append_units}} and \code{\link{resolve}}; may include unquoted column names}

\item{parse}{passed to \code{\link{append_units}}}
}
\value{
ggready
}
\description{
Prepares decorated data.frame for ggplot. Calls
\code{\link{resolve}} and appends
units to label using \code{\link{append_units}}
(passing \code{style = 'plotmath'} if \code{parse}
is true, else \code{style = 'plain'}).
Enforces classes 'decorated','resolved', and  'ggready'.
}
\examples{
example(ggready)
}
\seealso{
Other ggready: 
\code{\link{ggplot.ggready}()},
\code{\link{ggplot_build.ggready_ggplot}()},
\code{\link{ggready.data.frame}()},
\code{\link{ggready.resolved}()},
\code{\link{ggready}()},
\code{\link{print.ggready_ggplot}()}
}
\concept{ggready}
\keyword{internal}
