% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggready.R
\name{ggready}
\alias{ggready}
\title{Prepare Data for GGplot}
\usage{
ggready(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Prepares data for ggplot.
Generic, with method for data.frame.
}
\examples{
example(ggready.data.frame)
}
\seealso{
Other deprecated: 
\code{\link{alias.data.frame}()},
\code{\link{append_units}()},
\code{\link{ggready.data.frame}()}
}
\concept{deprecated}
\keyword{internal}
