% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{group_by_decorations.data.frame}
\alias{group_by_decorations.data.frame}
\title{Groups by Decorations for Data Frame}
\usage{
\method{group_by_decorations}{data.frame}(x, ...)
}
\arguments{
\item{x}{grouped_df}

\item{...}{ignored}
}
\value{
list of symbols
}
\description{
Invokes \code{\link[dplyr]{group_by}}
using whatever groups are recovered by
\code{\link{decorations_groups}}.
}
\examples{
library(magrittr)
library(dplyr)
Theoph \%>\% group_vars # nothing!
Theoph \%<>\% decorate_groups(Subject, Time) 
Theoph \%<>\% group_by_decorations
Theoph \%>\% group_vars # something
rm(Theoph)

}
\seealso{
Other decorate: 
\code{\link{as_decorated}()},
\code{\link{as_decorated.default}()},
\code{\link{decorate}()},
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{decorate.list}()},
\code{\link{decorate_groups}()},
\code{\link{decorate_groups.data.frame}()},
\code{\link{decorations}()},
\code{\link{decorations.data.frame}()},
\code{\link{decorations_groups}()},
\code{\link{decorations_groups.data.frame}()},
\code{\link{group_by_decorations}()},
\code{\link{redecorate}()}
}
\concept{decorate}
