% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{group_by_decorations}
\alias{group_by_decorations}
\title{Group by Decorations}
\usage{
group_by_decorations(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed}
}
\description{
Groups according to decorations.  Generic, 
with method \code{\link{group_by_decorations.data.frame}}
}
\seealso{
Other decorate: 
\code{\link{as_decorated}()},
\code{\link{as_decorated.default}()},
\code{\link{decorate}()},
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{decorate.list}()},
\code{\link{decorate_groups}()},
\code{\link{decorate_groups.data.frame}()},
\code{\link{decorations}()},
\code{\link{decorations.data.frame}()},
\code{\link{decorations_groups}()},
\code{\link{decorations_groups.data.frame}()},
\code{\link{group_by_decorations.data.frame}()},
\code{\link{redecorate}()}
}
\concept{decorate}
\keyword{internal}
