% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups_decorate.data.frame}
\alias{groups_decorate.data.frame}
\alias{groups_decorate}
\title{Capture Groups as Decorations for Data Frame}
\usage{
\method{groups_decorate}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{unquoted names of columns to assign as groups; defaults to \code{groups(x)}}
}
\value{
same class as x
}
\description{
Captures groups as decorations for class 'data.frame'.
Creates a sequentially-valued integer attribute 
with name 'groups' for each corresponding column
(after clearing all such existing designations).
It is an error if not all such columns are present.
Defaults to \code{groups(x)}. If no columns are 
specified and x has no groups, x is returned
with any existing column-level 'groups' attributes
removed.
}
\examples{
library(magrittr)
library(dplyr)
Theoph \%>\% groups_decorate(Subject, Time) \%>\% groups # nothing!
Theoph \%>\% groups_decorate(Subject, Time) \%>\% decorations # note well
Theoph \%>\% group_by(Subject, Time) \%>\% groups_decorate \%>\% decorations(Subject, Time) # same

}
\seealso{
Other decorate: 
\code{\link{as_decorated.default}()},
\code{\link{as_decorated}()},
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{decorate.list}()},
\code{\link{decorate}()},
\code{\link{decorations.data.frame}()},
\code{\link{decorations}()},
\code{\link{groups_decorate_generic}},
\code{\link{groups_decorations.data.frame}()},
\code{\link{groups_decorations_generic}},
\code{\link{redecorate}()}
}
\concept{decorate}
\keyword{internal}
