% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups_decorate}
\alias{groups_decorate_generic}
\title{Capture Groups as Decorations}
\usage{
groups_decorate(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed}
}
\description{
Captures groups as decorations.  Generic, 
with method \code{\link{groups_decorate.data.frame}}
}
\seealso{
Other decorate: 
\code{\link{as_decorated.default}()},
\code{\link{as_decorated}()},
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{decorate.list}()},
\code{\link{decorate}()},
\code{\link{decorations.data.frame}()},
\code{\link{decorations}()},
\code{\link{groups_decorate.data.frame}()},
\code{\link{groups_decorations.data.frame}()},
\code{\link{groups_decorations_generic}},
\code{\link{redecorate}()}
}
\concept{decorate}
\keyword{internal}
