% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups_decorations.data.frame}
\alias{groups_decorations.data.frame}
\alias{groups_decorations}
\title{Recover Groups Decorations for Data Frame}
\usage{
\method{groups_decorations}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{ignored}
}
\value{
list of symbols
}
\description{
Recovers groups decorations for class 'data.frame'.
Seeks a sequentially-valued integer attribute 
with name 'groups' for each column, sorts these,
and returns them as an un-named list of symbols,
like output of \code{groups(x)}.
}
\examples{
library(magrittr)
library(dplyr)
Theoph \%<>\% group_by(Subject, Time)
Theoph \%>\% groups_decorations # nothing!
Theoph \%<>\% groups_decorate
Theoph \%>\% groups_decorations # something!
Theoph \%<>\% ungroup
Theoph \%>\% groups # gone!
Theoph \%<>\% group_by(!!!groups_decorations(.))
Theoph \%>\% groups # recovered!


}
\seealso{
Other decorate: 
\code{\link{as_decorated.default}()},
\code{\link{as_decorated}()},
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{decorate.list}()},
\code{\link{decorate}()},
\code{\link{decorations.data.frame}()},
\code{\link{decorations}()},
\code{\link{groups_decorate.data.frame}()},
\code{\link{groups_decorate_generic}},
\code{\link{groups_decorations_generic}},
\code{\link{redecorate}()}
}
\concept{decorate}
\keyword{internal}
