% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explicit_guide.R
\name{implicit_guide}
\alias{implicit_guide}
\title{Coerce Guide to Something More Implicit}
\usage{
implicit_guide(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Coerces 'guide' to something more implicit.  Generic, with methods for
data.frame.  The key 'guide' generally suggests a guide
to interpretation of a data item, such as units, formats, codelists,
and encodings.  The idea here is to replace these with 'guide': i.e.,
to undo the effects of \code{\link{explicit_guide}}.
}
\seealso{
Other explicit_guide: 
\code{\link{explicit_guide.data.frame}()},
\code{\link{explicit_guide.yamlet}()},
\code{\link{explicit_guide}()},
\code{\link{implicit_guide.data.frame}()}
}
\concept{explicit_guide}
\keyword{internal}
