% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_yamlet.R
\name{io_yamlet.character}
\alias{io_yamlet.character}
\title{Import Yamlet}
\usage{
\method{io_yamlet}{character}(
  x,
  default_keys = getOption("yamlet_default_keys", list("label", "guide")),
  ...
)
}
\arguments{
\item{x}{file path for yamlet}

\item{default_keys}{character: default keys for the first n anonymous members of each element}

\item{...}{passed to \code{\link{as_yamlet}}}
}
\value{
yamlet: a named list with default keys applied
}
\description{
Imports yamlet.
Character method for \code{\link{io_yamlet}}.
Similar to \code{\link{read_yamlet}}, but only reads files.
}
\examples{
example(io_yamlet)
}
\seealso{
\code{\link{decorate.list}}

Other io: 
\code{\link{io_csv.character}()},
\code{\link{io_csv.data.frame}()},
\code{\link{io_csv}()},
\code{\link{io_table.character}()},
\code{\link{io_table.data.frame}()},
\code{\link{io_table}()},
\code{\link{io_yamlet.data.frame}()},
\code{\link{io_yamlet.yamlet}()},
\code{\link{io_yamlet}()}

Other interface: 
\code{\link{decorate.character}()},
\code{\link{decorate.list}()},
\code{\link{decorate}()},
\code{\link{io_csv.character}()},
\code{\link{io_csv.data.frame}()},
\code{\link{io_csv}()},
\code{\link{io_table.character}()},
\code{\link{io_table.data.frame}()},
\code{\link{io_table}()},
\code{\link{io_yamlet.data.frame}()},
\code{\link{io_yamlet.yamlet}()},
\code{\link{io_yamlet}()},
\code{\link{read_yamlet}()},
\code{\link{write_yamlet}()}
}
\concept{interface}
\concept{io}
