% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{is_dvec}
\alias{is_dvec}
\title{Test if Class is dvec}
\usage{
is_dvec(x)
}
\arguments{
\item{x}{object}
}
\value{
logical
}
\description{
Tests whether x inherits 'dvec'.
}
\examples{
is_dvec(1L)
is_dvec(as_dvec(1L))
}
