% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseable.R
\name{is_parseable}
\alias{is_parseable}
\title{Check Parseable as Units}
\usage{
is_parseable(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Checks if something is parseable as units.
Generic, with default method.
}
\examples{
example(is_parseable.character)
}
\seealso{
Other parseable: 
\code{\link{is_parseable.default}()}
}
\concept{parseable}
\keyword{internal}
