% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agplot.R
\name{isConditional.character}
\alias{isConditional.character}
\title{Test Character Expression is Conditional}
\usage{
\method{isConditional}{character}(x, data, ...)
}
\arguments{
\item{x}{default}

\item{data}{environment for variable lookup}

\item{...}{passed arguments}
}
\value{
logical
}
\description{
Tests whether character expression is conditional by default.
Evaluates x on data and looks for meaningful result.
}
\seealso{
Other conditional: 
\code{\link{isConditional.default}()},
\code{\link{isConditional}()}
}
\concept{conditional}
\keyword{internal}
