% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional.R
\name{isLevels.character}
\alias{isLevels.character}
\title{Test Character Value is Levels}
\usage{
\method{isLevels}{character}(x, table, ...)
}
\arguments{
\item{x}{default}

\item{table}{lookup vector}

\item{...}{passed arguments}
}
\value{
logical
}
\description{
Tests whether character value is levels.
Looks for any matches to vector.
Uses \code{\link{intersect}}, which is fairly flexible
respecting underlying data types (character 0 can match integer 0, etc.).
}
\seealso{
Other levels: 
\code{\link{isLevels.default}()},
\code{\link{isLevels}()}
}
\concept{levels}
\keyword{internal}
