% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional.R
\name{isLevels.default}
\alias{isLevels.default}
\title{Test Value is Levels by Default}
\usage{
\method{isLevels}{default}(x, table, ...)
}
\arguments{
\item{x}{default}

\item{...}{passed arguments}
}
\value{
logical
}
\description{
Tests whether value is levels by default.  Coerces to character.
}
\seealso{
Other levels: 
\code{\link{isLevels.character}()},
\code{\link{isLevels}()}
}
\concept{levels}
\keyword{internal}
