% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agplot.R
\name{isLevels}
\alias{isLevels}
\title{Test Value is Levels}
\usage{
isLevels(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{passed arguments}
}
\value{
logical
}
\description{
Tests whether value is levels.
}
\seealso{
Other levels: 
\code{\link{isLevels.character}()},
\code{\link{isLevels.default}()}
}
\concept{levels}
\keyword{internal}
