% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_label_units.R
\name{join_label_units.data.frame}
\alias{join_label_units.data.frame}
\title{Combine Labels and Units for Data Frame.}
\usage{
\method{join_label_units}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{passed to default method}
}
\value{
data.frame
}
\description{
Combines labels and units for data.frame.
For finer control, consider applying
\code{\link{join_label_units.default}}
to individual columns.
}
\examples{
library(magrittr)
library(ggplot2)
file <- system.file(package = 'yamlet', 'extdata','quinidine.csv')
file \%>\%
 decorate \%>\%
 filter(!is.na(conc)) \%>\%
 ggready(parse = FALSE) \%>\%
 ggplot(aes(x = time, y = conc, color = Heart)) +
 geom_point()

}
\seealso{
Other labels: 
\code{\link{express_label_units.data.frame}()},
\code{\link{express_label_units.default}()},
\code{\link{express_label_units}()},
\code{\link{join_label_units.default}()},
\code{\link{join_label_units}()}
}
\concept{labels}
