% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_label_units.R
\name{join_label_units.default}
\alias{join_label_units.default}
\title{Combine Label and Units By Default}
\usage{
\method{join_label_units}{default}(x, wrap = c(" (", ")"), ...)
}
\arguments{
\item{x}{object}

\item{wrap}{character to wrap units (will be coerced to length two)}

\item{...}{ignored arguments}
}
\value{
same class as x with updated label attribute
}
\description{
Combines label attribute and units attribute
as label attribute by default. Units attribute,
if present, is pasted to label.  If label is non-null,
units is wrapped in \code{wrap}.
}
\examples{
library(units)
library(magrittr)
x <- 1:10
attr(x, 'label') <- 'acceleration'
units(x) <- 'm/s^2'
y <- as_units('kg')
x \%>\% attr('label')
x \%>\% join_label_units \%>\% attr('label')
y \%>\% attr('label')
y \%>\% join_label_units \%>\% attr('label')


}
\seealso{
Other labels: 
\code{\link{express_label_units.data.frame}()},
\code{\link{express_label_units.default}()},
\code{\link{express_label_units}()},
\code{\link{join_label_units.data.frame}()},
\code{\link{join_label_units}()}
}
\concept{labels}
\keyword{internal}
