% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-to-latex.R
\name{latexToken}
\alias{latexToken}
\title{Process Latex Token}
\usage{
latexToken(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored arguments}
}
\value{
character
}
\description{
Pre-processes a latex token not recognized as
wikisymbol.  By default, simply escapes the
common upper and lower case names for Greek letters.
}
\examples{
latexToken('foo')
latexToken('alpha')
latexToken('Alpha')
}
\keyword{internal}
