% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{length.dvec}
\alias{length.dvec}
\title{Get Length of a Decorated Vector}
\usage{
\method{length}{dvec}(x)
}
\arguments{
\item{x}{decorated vector}
}
\value{
integer
}
\description{
Gets the length of a decorated vector.  Simply calls next method.
}
\examples{
length(as_dvec(1:3))
}
\seealso{
Other dvec: 
\code{\link{[.dvec}()},
\code{\link{[<-.dvec}()},
\code{\link{[[.dvec}()},
\code{\link{[[<-.dvec}()},
\code{\link{as_dvec.character}()},
\code{\link{as_dvec.integer}()},
\code{\link{as_dvec.logical}()},
\code{\link{as_dvec.numeric}()},
\code{\link{as_dvec}()},
\code{\link{c.dvec}()},
\code{\link{format.dvec}()},
\code{\link{print.dvec}()}
}
\concept{dvec}
\keyword{inernal}
