% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_striptext.R
\name{lines.striptext}
\alias{lines.striptext}
\title{Count Lines of Striptext}
\usage{
\method{lines}{striptext}(x, ...)
}
\arguments{
\item{x}{list with class 'striptext'}

\item{...}{passed arguments}
}
\value{
length one integer
}
\description{
Counts lines of striptext. Returns length two integer
indicating the highest number of lines encountered, per axis.
Line breaks are literal newlines, but also implied
where multiple columns exist per list element.
See also \code{\link{striptext}}.
}
\examples{
library(magrittr)
library(dplyr)
library(ggplot2)
(
  mtcars \%>\% 
  ggplot(aes(wt, mpg)) + 
  geom_point()
) \%>\% striptext \%>\% lines

(
  mtcars \%>\% 
  ggplot(aes(wt, mpg)) +
  geom_point() +
  facet_grid(cyl ~ am + gear)
) \%>\% striptext \%>\% lines

(
  mtcars \%>\% 
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(. ~ cyl + am + gear)
) \%>\% striptext \%>\% lines

(
  mtcars \%>\% 
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
    . ~ cyl + am + gear, 
    labeller = labeller(
      .default = label_both
    )
  )
) \%>\% striptext \%>\% lines

(
  mtcars \%>\% 
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
   . ~ cyl + am + gear, 
   labeller =  labeller(
     .default = label_both,
     .multi_line = FALSE
   )
 )
) \%>\% striptext \%>\% lines

(
  mtcars \%>\% 
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
   . ~ cyl + am + gear, 
   labeller = purrr::partial(
     label_both, 
     sep = "\n"
   )
 )
) \%>\% striptext \%>\% lines

(
  mtcars \%>\% 
  mutate(cyl = paste('cylinders:', cyl)) \%>\%
  mutate(gear = paste('gears:', gear)) \%>\%
  mutate(am = paste('transmission:', am))\%>\%
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
    . ~ cyl + am + gear, 
    labeller = labeller(
      .default = label_wrap_gen(10),
      .multi_line = TRUE
    )
  )
) \%>\% striptext \%>\% lines

(
  mtcars \%>\% 
  mutate(cyl =  paste(sep = '\n', 'cylinders', cyl)) \%>\%
  mutate(gear = paste(sep = '\n', 'gears', gear)) \%>\%
  mutate(am =   paste(sep = '\n', 'transmission', am))\%>\%
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
    . ~ cyl + am + gear, 
    labeller = labeller(
      .default = label_wrap_gen(10)
    )
  )
) \%>\% striptext \%>\% lines

(
  mtcars \%>\% 
  mutate(cyl =  paste(sep = ':', 'cylinders', cyl)) \%>\%
  mutate(gear = paste(sep = ':', 'gears', gear)) \%>\%
  mutate(am =   paste(sep = ':', 'transmission', am))\%>\%
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
    . ~ paste(sep = '\n', cyl, am, gear), 
  )
) \%>\% striptext \%>\% lines

}
\keyword{internal}
