% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{list2encoding}
\alias{list2encoding}
\title{Coerce List to Encoding}
\usage{
list2encoding(x, ...)
}
\arguments{
\item{x}{list}

\item{...}{ignored}
}
\value{
an encoding (length-one character), or original list if error occurs
}
\description{
Tries to coerce a list to an encoding.  Names are
understood as decodes, and list values as codes.
On failure, the list is returned unchanged.
}
\examples{
meta <- system.file(package = 'yamlet', 'extdata','quinidine.yaml')
meta <- as_yamlet(meta)
list2encoding(meta$Creatinine$guide)
}
\seealso{
Other encode: 
\code{\link{encode.yamlet}()}
}
\concept{encode}
\keyword{internal}
