% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_title.R
\name{make_title.decorated}
\alias{make_title.decorated}
\title{Make Title for Decorated}
\usage{
\method{make_title}{decorated}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{optional names of variables to limit scope}
}
\value{
decorated
}
\description{
Make title for class 'decorated'.
Limits scope to requested variables, and then calls
class-specific methods for each.
}
\examples{
library(magrittr)
x <- data.frame(length = 1:10)
x \%>\% 
  decorate('length: [ Length, mm ]') \%>\%
  resolve \%>\%
  decorations

}
\seealso{
Other labels: 
\code{\link{alias.data.frame}()},
\code{\link{append_units}()},
\code{\link{append_units.data.frame}()},
\code{\link{append_units.default}()},
\code{\link{drop_title}()},
\code{\link{drop_title.decorated}()},
\code{\link{drop_title.default}()},
\code{\link{drop_title.dvec}()},
\code{\link{make_title}()},
\code{\link{make_title.default}()},
\code{\link{make_title.dvec}()},
\code{\link{sub_units}()}
}
\concept{labels}
\keyword{internal}
