% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_title.R
\name{make_title.default}
\alias{make_title.default}
\title{Make Title by Default}
\usage{
\method{make_title}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\value{
same as x
}
\description{
Make title by default.
To be specific:  this is the default method
for the generic function \code{\link{make_title}},
and it actually does nothing.  Individual methods
are written for those classes where 'make title' 
behavior is expected.
}
\seealso{
Other labels: 
\code{\link{alias.data.frame}()},
\code{\link{append_units}()},
\code{\link{append_units.data.frame}()},
\code{\link{append_units.default}()},
\code{\link{drop_title}()},
\code{\link{drop_title.decorated}()},
\code{\link{drop_title.default}()},
\code{\link{drop_title.dvec}()},
\code{\link{make_title}()},
\code{\link{make_title.decorated}()},
\code{\link{make_title.dvec}()},
\code{\link{sub_units}()}
}
\concept{labels}
\keyword{internal}
