% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_title.R
\name{make_title}
\alias{make_title}
\title{Make Title}
\usage{
make_title(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Make title attribute.
Generic, with methods 
\code{\link{make_title.default}},
\code{\link{make_title.decorated}}, and
\code{\link{make_title.dvec}}.
}
\examples{
# see methods
}
\seealso{
Other deprecated: 
\code{\link{alias.data.frame}()},
\code{\link{append_units}()},
\code{\link{drop_title}()}

Other labels: 
\code{\link{alias.data.frame}()},
\code{\link{append_units.data.frame}()},
\code{\link{append_units.default}()},
\code{\link{append_units}()},
\code{\link{drop_title.decorated}()},
\code{\link{drop_title.default}()},
\code{\link{drop_title.dvec}()},
\code{\link{drop_title}()},
\code{\link{make_title.decorated}()},
\code{\link{make_title.default}()},
\code{\link{make_title.dvec}()},
\code{\link{sub_units}()}
}
\concept{deprecated}
\concept{labels}
\keyword{internal}
