% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapframe.R
\name{mapframe}
\alias{mapframe}
\title{Coerce Mapping to Dataframe}
\usage{
mapframe(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{ignored}
}
\value{
see methods
}
\description{
Coerces mapping to 'data.frame'.
Generic, with methods \code{\link{mapframe.default}}
and \code{\link{mapframe.ggplot}}.
}
\seealso{
Other decorated_ggplot: 
\code{\link{data_context}()},
\code{\link{data_context.ggplot}()},
\code{\link{ggplot.decorated}()},
\code{\link{ggplot_build.decorated_ggplot}()},
\code{\link{mapframe.default}()},
\code{\link{mapframe.ggplot}()}
}
\concept{decorated_ggplot}
\keyword{internal}
