% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge.decorated}
\alias{merge.decorated}
\title{Merge Decorated}
\usage{
\method{merge}{decorated}(x, y, ...)
}
\arguments{
\item{x}{decorated}

\item{y}{passed to \code{\link{merge}}}

\item{...}{passed to \code{\link{merge}}}
}
\value{
class 'decorated' 'data.frame'
}
\description{
Preserves class for 'decorated' during merge().
}
\examples{
library(magrittr)
library(dplyr)
x <- data.frame(foo = 1, bar = 2)
x \%<>\% decorate('foo: [distance, mm]')
x \%<>\% decorate('bar: [height, mm]')
class(merge(x,x))
}
\seealso{
Other decorated: 
\code{\link{[.decorated}()},
\code{\link{[<-.decorated}()},
\code{\link{[[.decorated}()},
\code{\link{[[<-.decorated}()},
\code{\link{names<-.decorated}()}
}
\concept{decorated}
