% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimic.R
\name{mimic.default}
\alias{mimic.default}
\title{Try To Look Like Another Equal-length Variable}
\usage{
\method{mimic}{default}(x, y = x, ...)
}
\arguments{
\item{x}{vector-like}

\item{y}{vector-like, same length as x}

\item{...}{ignored arguments}
}
\value{
same class as x
}
\description{
Tries to mimic another vector or factor.
If meaningful and possible, x acquires
a guide attribute with labels from
corresponding values in y. Any codelist
attribute is removed. No guide is created
for zero-length x.
}
\examples{
example(mimic.classified)

}
\seealso{
Other mimic: 
\code{\link{mimic.classified}()},
\code{\link{mimic}()}
}
\concept{mimic}
