% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimic.R
\name{mimic}
\alias{mimic}
\title{Try To Look Like Something Else}
\usage{
mimic(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Tries to make an object look like something else.
Generic, with method \code{\link{mimic.default}}
}
\examples{
example(mimic.default)
}
\seealso{
Other mimic: 
\code{\link{mimic.classified}()},
\code{\link{mimic.default}()}
}
\concept{mimic}
\keyword{internal}
