% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{modify.default}
\alias{modify.default}
\title{Modify Attributes of Indicated Components by Default}
\usage{
\method{modify}{default}(
  x,
  ...,
  .warn = getOption("yamlet_modify_warn", c("class", "levels", "labels", "names")),
  .silent = getOption("yamlet_modify_silent", TRUE)
)
}
\arguments{
\item{x}{object}

\item{...}{indicated columns, or name-value pairs}

\item{.warn}{reserved labels that warn on assignment}

\item{.silent}{passed to \code{\link{try}} (as 'silent')}
}
\value{
same class as x
}
\description{
Modifies the attributes of each indicated element
(all elements by default).  Assigns the value of an expression
to the supplied label, with existing attributes
and the object itself (.) available as arguments.
Fails silently on error. Gives a warning if the
supplied label is considered reserved.
}
\examples{
library(magrittr)
library(dplyr)
file <- system.file(package = 'yamlet', 'extdata','quinidine.csv')
x <- decorate(file)
x \%<>\% modify(.silent = FALSE, epithet = 2, time)
x \%<>\% modify(.silent = FALSE, epithet = 2 + 5, time)
x \%<>\% modify(.silent = TRUE, epithet = paste(label, '(', guide, ')'), time)
x \%<>\% modify(.silent = TRUE, epithet = paste(label, '(', guide, ')'), time)
x \%>\% select(time, conc) \%>\% as_yamlet
}
\seealso{
Other modify: 
\code{\link{modify}()}

Other interface: 
\code{\link{as_classified.factor}()},
\code{\link{conditionalize.data.frame}()},
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{ggplot.decorated}()},
\code{\link{ggready.data.frame}()},
\code{\link{io_csv.character}()},
\code{\link{io_csv.data.frame}()},
\code{\link{io_table.character}()},
\code{\link{io_table.data.frame}()},
\code{\link{io_yamlet.character}()},
\code{\link{io_yamlet.data.frame}()},
\code{\link{is_parseable.default}()},
\code{\link{read_yamlet}()},
\code{\link{resolve.data.frame}()},
\code{\link{write_yamlet}()}
}
\concept{interface}
\concept{modify}
