% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{modify}
\alias{modify}
\title{Modify Attributes of Indicated Components}
\usage{
modify(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Modifies the attributes of indicated components.
Generic, with default method.
}
\examples{
example(modify.default)
}
\seealso{
Other modify: 
\code{\link{modify.default}()},
\code{\link{named}()},
\code{\link{selected.default}()},
\code{\link{selected}()}
}
\concept{modify}
\keyword{internal}
