% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named.R
\name{named}
\alias{named}
\title{Evaluate Named Arguments}
\usage{
named(...)
}
\arguments{
\item{...}{possibly a mix of named and un-named arguments.}
}
\value{
named list
}
\description{
Evaluates named arguments in \code{...}.
}
\examples{
named(a = 1, b = 2 + 3, 4, 'd')

}
\seealso{
Other modify: 
\code{\link{modify.default}()},
\code{\link{modify}()},
\code{\link{selected.default}()},
\code{\link{selected}()}
}
\concept{modify}
\keyword{internal}
