% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonious.R
\name{parsimonious.default}
\alias{parsimonious.default}
\title{Reduce by Default to Simplest Sufficient Version}
\usage{
\method{parsimonious}{default}(x, ...)
}
\arguments{
\item{x}{object for which no specific parsimonious method exists}

\item{...}{ignored}
}
\description{
Reduces by default to simplest sufficient version.
This is a companion to \code{\link{parsimonious.list}}
and currently returns an unmodified object.
}
\seealso{
Other parsimonious: 
\code{\link{parsimonious.list}()},
\code{\link{parsimonious}()}
}
\concept{parsimonious}
\keyword{internal}
