% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonious.R
\name{parsimonious}
\alias{parsimonious}
\title{Reduce Something to its Simplest Sufficient Version}
\usage{
parsimonious(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{ignored}
}
\description{
Reduces something to its simplest sufficient version.
Generic, with method \code{\link{parsimonious.list}}.
}
\examples{
example(parsimonious.list)
}
\seealso{
Other parsimonious: 
\code{\link{parsimonious.default}()},
\code{\link{parsimonious.list}()}
}
\concept{parsimonious}
\keyword{internal}
