% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-to-plotmath.R
\name{plotmathToken}
\alias{plotmathToken}
\title{Quote a Token}
\usage{
plotmathToken(
  x,
  conditional = getOption("plotmath_conditional_quote", TRUE),
  unescape = getOption("plotmath_unescape", TRUE),
  ...
)
}
\arguments{
\item{x}{(length-one) character}

\item{conditional}{if true, return good tokens (parseable) unmodified}

\item{unescape}{whether to escape (unrecognized) backslash}

\item{...}{ignored arguments}
}
\value{
character
}
\description{
Quotes a token. Escapes single-quotes and wraps in single-quotes.
}
\examples{
plotmathToken("can't")
plotmathToken("\\\\$")
plotmathToken("\\\\$", unescape = FALSE)
plotmathToken("\\\\$") \%>\% goodToken
options(plotmath_unescape = FALSE)
plotmathToken("\\\\$") \%>\% goodToken
options(plotmath_unescape = NULL)
plotmathToken("\\\\$", unescape = FALSE) \%>\% goodToken
plotmathToken('foo')
plotmathToken('foo',conditional = FALSE)
options('plotmath_conditional_quote' = FALSE)
plotmathToken('foo')
options('plotmath_conditional_quote' = NULL)
plotmathToken('foo')
}
\keyword{internal}
