% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{print.decorations}
\alias{print.decorations}
\title{Print Decorations}
\usage{
\method{print}{decorations}(x, ...)
}
\arguments{
\item{x}{decorations, i.e. a named list of class 'decorations'}

\item{...}{ignored}
}
\value{
invisible x (yamlet)
}
\description{
Prints decorations.  Coerces to yamlet and prints result.
}
\examples{
example(decorations.data.frame)
}
\seealso{
Other decorate: 
\code{\link{as_decorated.default}()},
\code{\link{as_decorated}()},
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{decorate.list}()},
\code{\link{decorate}()},
\code{\link{decorations.data.frame}()},
\code{\link{decorations}()},
\code{\link{redecorate}()}
}
\concept{decorate}
