% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_ggready.R
\name{print.ggready_ggplot}
\alias{print.ggready_ggplot}
\title{Print Automatic Labels and Units for ggplot}
\usage{
\method{print}{ggready_ggplot}(x, ...)
}
\arguments{
\item{x}{class 'ggready_ggplot' from \code{\link{ggplot.ggready}}}

\item{...}{passed arguments}
}
\value{
see \code{\link[ggplot2]{print.ggplot}}
}
\description{
Prints automatic labels and units for ggplot.
Substitutes column label, if present, for default.
}
\examples{
example(ggready)
}
\seealso{
Other ggready: 
\code{\link{ggplot.ggready}()},
\code{\link{ggplot_build.ggready_ggplot}()},
\code{\link{ggready.data.frame}()},
\code{\link{ggready.decorated}()},
\code{\link{ggready}()}
}
\concept{ggready}
\keyword{internal}
