% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singularity.R
\name{promote.default}
\alias{promote.default}
\title{Promote by Default}
\usage{
\method{promote}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other arguments}
}
\description{
Promotes by default.  Currently a non-operation.
}
\keyword{internal}
