% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singularity.R
\name{promote}
\alias{promote}
\title{Promote Something.}
\usage{
promote(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Promotes something.  Generic, with default method.
}
\examples{
example(promote.data.frame)
}
\seealso{
Other promote: 
\code{\link{[.decorated}()},
\code{\link{filter.decorated}()},
\code{\link{promote.default}()},
\code{\link{singularity}()}
}
\concept{promote}
\keyword{internal}
