% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{proxy}
\alias{proxy}
\title{Calculate Substitute Values}
\usage{
proxy(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Calculates substitute values.
Generic, with method \code{\link{proxy.factor}}
}
\examples{
example(proxy.factor)
}
\seealso{
Other proxy: 
\code{\link{proxy.factor}()}
}
\concept{proxy}
\keyword{internal}
