% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-to-plotmath.R
\name{quoteToken}
\alias{quoteToken}
\title{Quote a Token}
\usage{
quoteToken(
  x,
  conditional = getOption("plotmath_conditional_quote", TRUE),
  unescape = getOption("plotmath_unescape", TRUE),
  ...
)
}
\arguments{
\item{x}{(length-one) character}

\item{conditional}{if true, return good tokens (parseable) unmodified}

\item{unescape}{whether to escape (unrecognized) backslash}

\item{...}{ignored arguments}
}
\value{
character
}
\description{
Quotes a token. Escapes single-quotes and wraps in single-quotes.
}
\examples{
quoteToken("can't")
quoteToken("\\\\$")
quoteToken("\\\\$", unescape = FALSE)
quoteToken("\\\\$") \%>\% goodToken
options(plotmath_unescape = FALSE)
quoteToken("\\\\$") \%>\% goodToken
options(plotmath_unescape = NULL)
quoteToken("\\\\$", unescape = FALSE) \%>\% goodToken
quoteToken('foo')
quoteToken('foo',conditional = FALSE)
options('plotmath_conditional_quote' = FALSE)
quoteToken('foo')
options('plotmath_conditional_quote' = NULL)
quoteToken('foo')
}
\keyword{internal}
