% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{reconcile.list}
\alias{reconcile.list}
\title{Reconcile List Atttributes}
\usage{
\method{reconcile}{list}(x, ...)
}
\arguments{
\item{x}{list}

\item{...}{passed arguments}
}
\value{
list (of attributes)
}
\description{
Reconciles list attributes. Recursively arbitrates
list members pairwise, returning the accumulated result.
}
\keyword{internal}
