% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{reconcile}
\alias{reconcile}
\title{Reconcile Atttributes}
\usage{
reconcile(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\value{
list (of attributes)
}
\description{
Reconciles attributes.  Generic, with method \code{\link{reconcile.list}}.
}
\keyword{internal}
