% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_attr.R
\name{render_attr.data.frame}
\alias{render_attr.data.frame}
\title{Render Attribute for Data Frame}
\usage{
\method{render_attr}{data.frame}(x, which, fun = concatenate, ...)
}
\arguments{
\item{x}{data.frame}

\item{which}{attribute name (exact match); see \code{\link{attr}}}

\item{fun}{function or function name to process attribute}

\item{...}{passed arguments}
}
\value{
data.frame
}
\description{
Renders Attribute for data.frame.
Calls \code{\link{render_attr}}
for each column.
}
\examples{
library(magrittr)
file <- system.file(package = 'yamlet', 'extdata','quinidine.csv')
file \%>\%
decorate \%>\%
explicit_guide \%>\%
append_units \%>\%
render_attr('label', toupper) \%>\%
as_yamlet
}
