% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_attr.R
\name{render_attr.default}
\alias{render_attr.default}
\title{Render Attribute by Default}
\usage{
\method{render_attr}{default}(x, which, fun = concatenate, ...)
}
\arguments{
\item{x}{object}

\item{which}{attribute name (exact match); see \code{\link{attr}}}

\item{fun}{function or function name to process attribute}

\item{...}{passed arguments}
}
\value{
same class as x with updated attribute
}
\description{
Renders Attribute by default.
}
\examples{
library(magrittr)
file <- system.file(package = 'yamlet', 'extdata','quinidine.csv')
x <- file \%>\% decorate
attributes(x$time)
render_attr(x$time, 'label', toupper) \%>\% attributes
}
