% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_attr.R
\name{render_attr}
\alias{render_attr}
\title{Render Attribute}
\usage{
render_attr(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Renders Attribute
Generic, with methods \code{\link{render_attr.default}}
and \code{\link{render_attr.data.frame}}.
}
\examples{
# see methods
}
\keyword{internal}
