% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{rescale.dvec}
\alias{rescale.dvec}
\title{Rescale dvec}
\usage{
\method{rescale}{dvec}(x, to = c(0, 1), from = range(x, na.rm = TRUE, finite = TRUE), ...)
}
\arguments{
\item{x}{dvec}

\item{to}{numeric}

\item{from}{numeric}

\item{...}{passed arguments}
}
\description{
Rescales dvec
}
\keyword{internal}
