% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_striptext.R
\name{scale_striptext}
\alias{scale_striptext}
\title{Scale Striptext}
\usage{
scale_striptext(object = getOption("ggplot_scale_striptext", 0.5), ...)
}
\arguments{
\item{object}{ggplot}

\item{...}{ignored}

\item{scale}{length-one numeric on [0, 1]}
}
\value{
ggplot_scale_striptext
}
\description{
Scales striptext font size by number of striptext lines.
Relative font size is adjusted on a per-axis
basis using \code{(1/x)^n}, where 
\code{x} is the number of striptext lines and
\code{n} is in the interval [0, 1]. 
\code{n} is 0.5 by default but can be passed as the argument \code{object}:
e.g. \code{plot <- plot + scale_striptext(1)}.
}
\examples{
library(magrittr)
library(dplyr)
library(ggplot2)

  mtcars \%>\% 
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  scale_striptext()

  mtcars \%>\% 
  ggplot(aes(wt, mpg)) +
  geom_point() +
  facet_grid(cyl ~ am + gear) +
  scale_striptext()

  mtcars \%>\% 
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(. ~ cyl + am + gear) +
  scale_striptext()

  mtcars \%>\% 
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
    . ~ cyl + am + gear, 
    labeller = labeller(
      .default = label_both
    )
  ) +
  scale_striptext()

  mtcars \%>\% 
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
    . ~ cyl + am + gear, 
    labeller =  labeller(
      .default = label_both,
      .multi_line = FALSE
    )
  ) +
  scale_striptext()

  mtcars \%>\% 
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
    . ~ cyl + am + gear, 
    labeller = purrr::partial(
      label_both, 
      sep = "\n"
    )
  ) +
  scale_striptext()

p <- mtcars \%>\% 
  mutate(cyl = paste('cylinders:', cyl)) \%>\%
  mutate(gear = paste('gears:', gear)) \%>\%
  mutate(am = paste('transmission:', am))\%>\%
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
    . ~ cyl + am + gear, 
    labeller = labeller(
      .default = label_wrap_gen(10)
    )
  ) 
  
  p + scale_striptext(0)
  p + scale_striptext(0.5)
  p + scale_striptext(1)

  mtcars \%>\% 
  mutate(cyl =  paste(sep = '\n', 'cylinders', cyl)) \%>\%
  mutate(gear = paste(sep = '\n', 'gears', gear)) \%>\%
  mutate(am =   paste(sep = '\n', 'transmission', am))\%>\%
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
    . ~ cyl + am + gear, 
    labeller = labeller(
      .default = label_wrap_gen(10)
    )
  ) + 
  scale_striptext()

  mtcars \%>\% 
  mutate(cyl =  paste(sep = ':', 'cylinders', cyl)) \%>\%
  mutate(gear = paste(sep = ':', 'gears', gear)) \%>\%
  mutate(am =   paste(sep = ':', 'transmission', am))\%>\%
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
    . ~ paste(sep = '\n', cyl, am, gear), 
  ) + 
  scale_striptext()

}
\seealso{
ggplot_add.ggplot_scale_striptext

Other isometric: 
\code{\link{ggplot_add.ggplot_isometric}()},
\code{\link{ggplot_add.ggplot_scale_striptext}()},
\code{\link{ggplot_add.ggplot_symmetric}()},
\code{\link{isometric}()},
\code{\link{symmetric}()}
}
\concept{isometric}
\keyword{internal}
