% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{scale_type.dvec}
\alias{scale_type.dvec}
\title{Determine Scale Type for dvec}
\usage{
\method{scale_type}{dvec}(x)
}
\arguments{
\item{x}{dvec}
}
\description{
Determines scale type for dvec.
}
\keyword{internal}
