% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected.R
\name{selected}
\alias{selected}
\title{Identify Selected Names}
\usage{
selected(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Identifies Selected Names.
Generic, with default method.
}
\examples{
example(modify.data.frame)
}
\seealso{
Other modify: 
\code{\link{modify.default}()},
\code{\link{modify}()},
\code{\link{named}()},
\code{\link{selected.default}()}
}
\concept{modify}
\keyword{internal}
