% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{slice.decorated}
\alias{slice.decorated}
\title{Slice Decorated}
\usage{
\method{slice}{decorated}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{A tbl.}

\item{...}{Integer row values.  Provide either positive values to keep,
or negative values to drop. The values provided must be either all
positive or all negative.  Indices beyond the number of rows in the
input are silently ignored.

The arguments in \code{...} are automatically \link[rlang:quo]{quoted} and
\link[rlang:eval_tidy]{evaluated} in the context of the data
frame. They support \link[rlang:quasiquotation]{unquoting} and
splicing. See \code{vignette("programming")} for an introduction to
these concepts.}

\item{.preserve}{when \code{FALSE} (the default), the grouping structure
is recalculated based on the resulting data, otherwise it is kept as is.}
}
\description{
Preserves class when slicing decorated.
}
\seealso{
Other dplyr: 
\code{\link{anti_join.decorated}()},
\code{\link{arrange.decorated}()},
\code{\link{as_decorated}()},
\code{\link{filter.decorated}()},
\code{\link{full_join.decorated}()},
\code{\link{group_by.decorated}()},
\code{\link{inner_join.decorated}()},
\code{\link{left_join.decorated}()},
\code{\link{mutate.decorated}()},
\code{\link{right_join.decorated}()},
\code{\link{select.decorated}()},
\code{\link{semi_join.decorated}()},
\code{\link{summarise.decorated}()},
\code{\link{summarize.decorated}()}
}
\concept{dplyr}
\keyword{internal}
