% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/striptext.R
\name{striptext.ggplot}
\alias{striptext.ggplot}
\title{Extract Striptext from ggplot}
\usage{
\method{striptext}{ggplot}(x, ...)
}
\arguments{
\item{x}{ggplot}

\item{...}{ignored}
}
\value{
list
}
\description{
Extracts striptext from ggplot.
Generic, with method \code{\link{striptext.ggplot}}.
}
\examples{
example(striptext)
}
\keyword{internal}
