% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/striptext.R
\name{striptext}
\alias{striptext}
\title{Extract Striptext}
\usage{
striptext(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\value{
list with class 'striptext'
}
\description{
Extracts striptext.
Generic, with method \code{\link{striptext.ggplot}}.
}
\examples{
library(magrittr)
library(dplyr)
library(ggplot2)
(
  mtcars \%>\% 
  ggplot(aes(wt, mpg)) + 
  geom_point()
) \%>\% striptext

(
  mtcars \%>\% 
  ggplot(aes(wt, mpg)) +
  geom_point() +
  facet_grid(cyl ~ am + gear)
) \%>\% striptext

(
  mtcars \%>\% 
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(. ~ cyl + am + gear)
) \%>\% striptext

(
  mtcars \%>\% 
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
    . ~ cyl + am + gear, 
    labeller = labeller(
      .default = label_both
    )
  )
) \%>\% striptext

(
  mtcars \%>\% 
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
   . ~ cyl + am + gear, 
   labeller =  labeller(
     .default = label_both,
     .multi_line = FALSE
   )
 )
) \%>\% striptext

(
  mtcars \%>\% 
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
   . ~ cyl + am + gear, 
   labeller = purrr::partial(
     label_both, 
     sep = "\n"
   )
 )
) \%>\% striptext

(
  mtcars \%>\% 
  mutate(cyl = paste('cylinders:', cyl)) \%>\%
  mutate(gear = paste('gears:', gear)) \%>\%
  mutate(am = paste('transmission:', am))\%>\%
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
    . ~ cyl + am + gear, 
    labeller = labeller(
      .default = label_wrap_gen(10),
      .multi_line = TRUE
    )
  )
) \%>\% striptext

(
  mtcars \%>\% 
  mutate(cyl =  paste(sep = '\n', 'cylinders', cyl)) \%>\%
  mutate(gear = paste(sep = '\n', 'gears', gear)) \%>\%
  mutate(am =   paste(sep = '\n', 'transmission', am))\%>\%
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
    . ~ cyl + am + gear, 
    labeller = labeller(
      .default = label_wrap_gen(10)
    )
  )
) \%>\% striptext

(
  mtcars \%>\% 
  mutate(cyl =  paste(sep = ':', 'cylinders', cyl)) \%>\%
  mutate(gear = paste(sep = ':', 'gears', gear)) \%>\%
  mutate(am =   paste(sep = ':', 'transmission', am))\%>\%
  ggplot(aes(wt, mpg)) + 
  geom_point() +
  facet_wrap(
    . ~ paste(sep = '\n', cyl, am, gear), 
  )
) \%>\% striptext

}
\keyword{internal}
