% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified.R
\name{[.classified}
\alias{[.classified}
\title{Subset Classified.}
\usage{
\method{[}{classified}(x, ...)
}
\arguments{
\item{x}{classified factor}

\item{...}{passed to next method}
}
\value{
class 'classified' 'factor'
}
\description{
Subsets classified factor, retaining attributes.
}
\examples{
a <- as_classified(factor(letters))
attr(a, 'label') <- 'foo'
a <- a[1:3]
attributes(a)
}
\seealso{
Other classified: 
\code{\link{as_classified.factor}()},
\code{\link{as_classified}()}
}
\concept{classified}
\keyword{internal}
