% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{[.dvec}
\alias{[.dvec}
\title{Subset Decorated Vector}
\usage{
\method{[}{dvec}(x, ...)
}
\arguments{
\item{x}{decorated vector}

\item{...}{passed to next method}
}
\value{
dvec
}
\description{
Subsets decorated vector, retaining attributes.
}
\examples{
a <- as_dvec(letters[1:3])
attr(a, 'label') <- 'foo'
a <- a[1:3]
attributes(a)
}
\seealso{
Other dvec: 
\code{\link{[<-.dvec}()},
\code{\link{[[.dvec}()},
\code{\link{[[<-.dvec}()},
\code{\link{as_dvec.character}()},
\code{\link{as_dvec.integer}()},
\code{\link{as_dvec.logical}()},
\code{\link{as_dvec.numeric}()},
\code{\link{as_dvec}()},
\code{\link{c.dvec}()},
\code{\link{format.dvec}()},
\code{\link{length.dvec}()},
\code{\link{print.dvec}()}
}
\concept{dvec}
\keyword{internal}
