% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_string.R
\name{[.unit_string}
\alias{[.unit_string}
\title{Subset Unit String}
\usage{
\method{[}{unit_string}(x, ...)
}
\arguments{
\item{x}{unit_string}

\item{...}{passed to next method}
}
\value{
unit_string
}
\description{
Subsets unit_string, retaining class.
}
\examples{
x <- c(
  'm',
  's',
  'ng/mL'
)
x <- as_unit_string(x)
class(x)
class(x[1])
}
\seealso{
Other unit_string: 
\code{\link{[[.unit_string}()},
\code{\link{as_unit_string.character}()},
\code{\link{as_unit_string.factor}()},
\code{\link{as_unit_string.symbolic_units}()},
\code{\link{as_unit_string.units}()},
\code{\link{as_unit_string}()}
}
\concept{unit_string}
\keyword{internal}
