% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{[.yamlet}
\alias{[.yamlet}
\title{Subset Yamlet}
\usage{
\method{[}{yamlet}(x, ...)
}
\arguments{
\item{x}{object to subset}

\item{...}{passed to next method}
}
\value{
yamlet
}
\description{
Subsets yamlet. Preserves class, since a subset of yamlet is still yamlet.
}
\examples{
meta <- system.file(package = 'yamlet', 'extdata','yamlet.yaml')
meta <- as_yamlet(meta)
class(meta)
stopifnot(inherits(meta[1:2],'yamlet'))
}
\seealso{
Other encode: 
\code{\link{encode.yamlet}()},
\code{\link{list2encoding}()}
}
\concept{encode}
\keyword{internal}
