% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified.R
\name{[[.classified}
\alias{[[.classified}
\title{Element-select Classified}
\usage{
\method{[[}{classified}(x, ...)
}
\arguments{
\item{x}{classified factor}

\item{...}{passed to next method}
}
\value{
class 'classified' 'factor'
}
\description{
Selects element of classified factor, retaining attributes.
}
\examples{
a <- classified(letters[1:3])
attr(a, 'label') <- 'foo'
a <- a[[2]]
attributes(a)
}
\seealso{
Other classified: 
\code{\link{[.classified}()},
\code{\link{[<-.classified}()},
\code{\link{[[<-.classified}()},
\code{\link{as.integer.classified}()},
\code{\link{c.classified}()},
\code{\link{classified.data.frame}()},
\code{\link{classified.default}()},
\code{\link{classified}()},
\code{\link{unclassified.classified}()},
\code{\link{unclassified.data.frame}()},
\code{\link{unclassified}()}
}
\concept{classified}
\keyword{internal}
