% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{[[.plotmath}
\alias{[[.plotmath}
\title{Element-select Plotmath}
\usage{
\method{[[}{plotmath}(x, ...)
}
\arguments{
\item{x}{plotmath}

\item{...}{passed to next method}
}
\value{
plotmath
}
\description{
Element-selects plotmath, retaining class.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_plotmath(as_wikisymbol(x))
class(x)
class(x[[1]])
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol.symbolic_units}()},
\code{\link{as_wikisymbol}()},
\code{\link{wikisym2latex_}()},
\code{\link{wikisym2plotmath_}()}
}
\concept{wikisymbol}
\keyword{internal}
