% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{[[.wikisymbol}
\alias{[[.wikisymbol}
\title{Element-select Wiki Symbol}
\usage{
\method{[[}{wikisymbol}(x, ...)
}
\arguments{
\item{x}{wikisymbol}

\item{...}{passed to next method}
}
\value{
wikisymbol
}
\description{
Element-selects wikisymbol, retaining class.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_wikisymbol(x)
class(x)
class(x[[1]])
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol.symbolic_units}()},
\code{\link{as_wikisymbol}()},
\code{\link{wiki_to_latex}()},
\code{\link{wiki_to_plotmath}()}
}
\concept{wikisymbol}
\keyword{internal}
