% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorated.R
\name{[<-.decorated}
\alias{[<-.decorated}
\title{Assign Subset of Decorated}
\usage{
\method{[}{decorated}(x, ...) <- value
}
\arguments{
\item{x}{decorated}

\item{...}{passed to next method}
}
\value{
decorated
}
\description{
Assigns subset of decorated, retaining attributes.
}
\examples{
a <- as_decorated(as.list(setNames(letters[1:3], LETTERS[1:3])))
a[2:3] <- 'a'
str(a)
class(a)
}
\seealso{
Other decorated: 
\code{\link{[.decorated}()},
\code{\link{[[.decorated}()},
\code{\link{[[<-.decorated}()},
\code{\link{merge.decorated}()},
\code{\link{names<-.decorated}()}
}
\concept{decorated}
\keyword{internal}
