% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{to_yamlet.default}
\alias{to_yamlet.default}
\title{Coerce Default to Yamlet Storage Format}
\usage{
\method{to_yamlet}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\value{
length-one character
}
\description{
Coerces to yamlet storage format by default conversion to character.
}
\examples{
to_yamlet(3)
to_yamlet(c(a = '4',b = '5.8'))
to_yamlet(c(a = 4,b = 5.8))
to_yamlet(TRUE)
}
\seealso{
Other to_yamlet: 
\code{\link{to_yamlet.NULL}()},
\code{\link{to_yamlet.character}()},
\code{\link{to_yamlet.list}()},
\code{\link{to_yamlet.yamlet}()},
\code{\link{to_yamlet}()}
}
\concept{to_yamlet}
\keyword{internal}
