% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{to_yamlet.NULL}
\alias{to_yamlet.NULL}
\title{Coerce Null to Yamlet Storage Format}
\usage{
\method{to_yamlet}{`NULL`}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\value{
length-one character
}
\description{
Coerces null to yamlet storage format (returns empty string).
}
\examples{
to_yamlet(NULL)
}
\seealso{
Other to_yamlet: 
\code{\link{to_yamlet.character}()},
\code{\link{to_yamlet.default}()},
\code{\link{to_yamlet.list}()},
\code{\link{to_yamlet.yamlet}()},
\code{\link{to_yamlet}()}
}
\concept{to_yamlet}
\keyword{internal}
