% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{to_yamlet}
\alias{to_yamlet}
\title{Coerce to Yamlet Storage Format}
\usage{
to_yamlet(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\value{
length-one character
}
\description{
Coerces to yamlet storage format. Generic, with methods
for default, null, character and list which together
implement the yamlet storage syntax.
Always returns length-one character, possibly the empty string.
}
\seealso{
Other to_yamlet: 
\code{\link{to_yamlet.NULL}()},
\code{\link{to_yamlet.character}()},
\code{\link{to_yamlet.default}()},
\code{\link{to_yamlet.list}()}
}
\concept{to_yamlet}
\keyword{internal}
