% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{to_yamlet.yamlet}
\alias{to_yamlet.yamlet}
\title{Coerce Yamlet to Yamlet Storage Format}
\usage{
\method{to_yamlet}{yamlet}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\value{
length-one character
}
\description{
Coerces yamlet to yamlet storage format by unclassing to list.
}
\examples{
library(magrittr)
 'a: [[d: [0, 1, 2]]]' \%>\% as_yamlet \%>\% to_yamlet
}
\seealso{
Other to_yamlet: 
\code{\link{to_yamlet.NULL}()},
\code{\link{to_yamlet.character}()},
\code{\link{to_yamlet.default}()},
\code{\link{to_yamlet.list}()},
\code{\link{to_yamlet}()}
}
\concept{to_yamlet}
\keyword{internal}
