% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified.R
\name{type_sum.classified}
\alias{type_sum.classified}
\title{Summarize Type of Classified}
\usage{
\method{type_sum}{classified}(x)
}
\arguments{
\item{x}{classified}
}
\description{
Summarizes type of classified.
}
\examples{
type_sum(classified(0))
}
\keyword{internal}
