% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclassified.R
\name{unclassified.dvec}
\alias{unclassified.dvec}
\title{Unclassify Decorated Vector}
\usage{
\method{unclassified}{dvec}(x, ...)
}
\arguments{
\item{x}{dvec}

\item{...}{ignored}
}
\value{
dvec
}
\description{
Unclassifies dvec.  A non-operation, since dvec is not classified.
Needed for completness by \code{\link{resolve.dvec}}.
}
\examples{
library(magrittr)
x <- structure(as_dvec(1), guide = 'misc')
unclassified(x)
}
\seealso{
Other dvec: 
\code{\link{[.dvec}()},
\code{\link{[<-.dvec}()},
\code{\link{[[.dvec}()},
\code{\link{[[<-.dvec}()},
\code{\link{as.data.frame.dvec}()},
\code{\link{as_dvec}()},
\code{\link{as_dvec.character}()},
\code{\link{as_dvec.complex}()},
\code{\link{as_dvec.dvec}()},
\code{\link{as_dvec.integer}()},
\code{\link{as_dvec.logical}()},
\code{\link{as_dvec.numeric}()},
\code{\link{c.dvec}()},
\code{\link{classified.dvec}()},
\code{\link{desolve.dvec}()},
\code{\link{explicit_guide.data.frame}()},
\code{\link{format.dvec}()},
\code{\link{implicit_guide.dvec}()},
\code{\link{length.dvec}()},
\code{\link{print.dvec}()},
\code{\link{resolve.classified}()},
\code{\link{resolve.dvec}()}
}
\concept{dvec}
\keyword{internal}
