% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclassified.R
\name{unclassified}
\alias{unclassified}
\title{Unclassify Something}
\usage{
unclassified(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Unclassify something.
Generic, with method \code{\link{unclassified.classified}}.
See also \code{\link{classified}}.
}
\examples{
example(unclassified.classified)
}
\seealso{
Other classified: 
\code{\link{[.classified}()},
\code{\link{[<-.classified}()},
\code{\link{[[.classified}()},
\code{\link{[[<-.classified}()},
\code{\link{as.integer.classified}()},
\code{\link{c.classified}()},
\code{\link{classified.data.frame}()},
\code{\link{classified.default}()},
\code{\link{classified}()},
\code{\link{unclassified.classified}()},
\code{\link{unclassified.data.frame}()}
}
\concept{classified}
\keyword{internal}
