% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{undecorate.default}
\alias{undecorate.default}
\title{Undecorate by Default}
\usage{
\method{undecorate}{default}(x, as.is = TRUE, ...)
}
\arguments{
\item{x}{object}

\item{as.is}{passed to \code{\link{type.convert}}}

\item{...}{passed arguments}
}
\value{
a list-like object, typically data.frame
}
\description{
Undecorates by default method. Calls \code{\link{type.convert}} to 
each element, with \code{as.is = TRUE} by default.
}
\examples{
file <- system.file(package = 'yamlet', 'extdata','xanomeline.csv.gz')
x <- io_csv(file)
head(decorations(x))
head(decorations(undecorate(x)))
}
\seealso{
Other decorate: 
\code{\link{as_decorated}()},
\code{\link{as_decorated.default}()},
\code{\link{decorate}()},
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{decorate.list}()},
\code{\link{decorate_groups}()},
\code{\link{decorate_groups.data.frame}()},
\code{\link{decorations}()},
\code{\link{decorations.data.frame}()},
\code{\link{decorations_groups}()},
\code{\link{decorations_groups.data.frame}()},
\code{\link{group_by_decorations}()},
\code{\link{group_by_decorations.data.frame}()},
\code{\link{redecorate}()},
\code{\link{undecorate}()}
}
\concept{decorate}
