% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{undecorate}
\alias{undecorate}
\title{Undecorate Something}
\usage{
undecorate(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
a list-like object, typically data.frame
}
\description{
Undecorates something.
Generic, with method \code{\link{undecorate.default}}.
}
\seealso{
Other decorate: 
\code{\link{as_decorated}()},
\code{\link{as_decorated.default}()},
\code{\link{decorate}()},
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{decorate.list}()},
\code{\link{decorate_groups}()},
\code{\link{decorate_groups.data.frame}()},
\code{\link{decorations}()},
\code{\link{decorations.data.frame}()},
\code{\link{decorations_groups}()},
\code{\link{decorations_groups.data.frame}()},
\code{\link{group_by_decorations}()},
\code{\link{group_by_decorations.data.frame}()},
\code{\link{redecorate}()},
\code{\link{undecorate.default}()}
}
\concept{decorate}
\keyword{internal}
