% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{unnest.list}
\alias{unnest.list}
\title{Collapse Uninformative Levels of a List}
\usage{
\method{unnest}{list}(x, ...)
}
\arguments{
\item{x}{list}
}
\value{
list
}
\description{
The list method for unnest() recursively
ascends a nested list, removing uninformative levels.
}
\examples{
unnest(yaml::yaml.load('ITEM: [ label: sunshine, [foo: 1, bar: 3]]'))
as_yamlet('ITEM: [ label: sunshine, [foo: 1, bar: 3]]')
}
\seealso{
Other unnest: 
\code{\link{unnest.default}()},
\code{\link{unnest}()}
}
\concept{unnest}
\keyword{internal}
