% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_cast.R
\name{vec_cast.character.dvec}
\alias{vec_cast.character.dvec}
\title{Cast to character from dvec}
\usage{
\method{vec_cast}{character.dvec}(x, to, ...)
}
\arguments{
\item{to}{character}

\item{...}{ignored}
}
\description{
Cast to character from dvec.
}
\keyword{internal}
