% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_classified.R
\name{vec_cast.classified.classified}
\alias{vec_cast.classified.classified}
\title{Cast to classified from classified}
\usage{
\method{vec_cast}{classified.classified}(x, to, ...)
}
\arguments{
\item{x}{classified}

\item{to}{classified}

\item{...}{ignored}
}
\description{
Cast to classified from classified.
}
\keyword{internal}
