% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_classified.R
\name{vec_cast.classified.factor}
\alias{vec_cast.classified.factor}
\title{Cast to classified from factor}
\usage{
\method{vec_cast}{classified.factor}(x, to, ...)
}
\arguments{
\item{x}{factor}

\item{to}{classified}

\item{...}{ignored}
}
\description{
Cast to classified from factor
}
\keyword{internal}
