% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_cast.R
\name{vec_cast.complex.dvec}
\alias{vec_cast.complex.dvec}
\title{Cast to complex from dvec}
\usage{
\method{vec_cast}{complex.dvec}(x, to, ...)
}
\arguments{
\item{to}{complex}

\item{...}{ignored}
}
\description{
Cast to complex from dvec.
}
\keyword{internal}
