% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_decorated.R
\name{vec_cast.data.frame.decorated}
\alias{vec_cast.data.frame.decorated}
\title{Cast to Data.frame from Decorated}
\usage{
\method{vec_cast}{data.frame.decorated}(x, to, ...)
}
\arguments{
\item{x}{decorated}

\item{to}{data.frame}

\item{...}{passed arguments}
}
\value{
data.frame
}
\description{
Casts to data.frame from decorated.
}
