% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_decorated.R
\name{vec_cast.decorated.data.frame}
\alias{vec_cast.decorated.data.frame}
\title{Cast to Decorated from Data.frame}
\usage{
\method{vec_cast}{decorated.data.frame}(x, to, ...)
}
\arguments{
\item{x}{data.frame}

\item{to}{decorated}

\item{...}{passed arguments}
}
\value{
decorated
}
\description{
Casts to decorated from data.frame.
}
