% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_decorated.R
\name{vec_cast.decorated.decorated}
\alias{vec_cast.decorated.decorated}
\title{Cast to Decorated from Decorated}
\usage{
\method{vec_cast}{decorated.decorated}(x, to, ...)
}
\arguments{
\item{x}{decorated}

\item{to}{decorated}

\item{...}{passed arguments}
}
\value{
decorated
}
\description{
Casts to decorated from decorated.
}
\keyword{internal}
