% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_cast.R
\name{vec_cast.double.dvec}
\alias{vec_cast.double.dvec}
\title{Cast to double from dvec}
\usage{
\method{vec_cast}{double.dvec}(x, to, ...)
}
\arguments{
\item{to}{double}

\item{...}{passed arguments}

\item{from}{dvec}
}
\description{
Cast to double from dvec.
}
\keyword{internal}
