% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_cast.R
\name{vec_cast.dvec.character}
\alias{vec_cast.dvec.character}
\title{Cast to dvec from character}
\usage{
\method{vec_cast}{dvec.character}(x, to, ...)
}
\arguments{
\item{to}{dvec}

\item{...}{passed arguments}

\item{from}{character}
}
\description{
Cast to dvec from character.
}
\keyword{internal}
