% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_cast.R
\name{vec_cast.dvec.complex}
\alias{vec_cast.dvec.complex}
\title{Cast to dvec from complex}
\usage{
\method{vec_cast}{dvec.complex}(x, to, ...)
}
\arguments{
\item{to}{dvec}

\item{...}{ignored}

\item{from}{complex}
}
\description{
Cast to dvec from complex.
}
\keyword{internal}
