% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_cast.R, R/vec_cast_dvec.R
\name{vec_cast.dvec.double}
\alias{vec_cast.dvec.double}
\title{Cast to dvec from double}
\usage{
\method{vec_cast}{dvec.double}(x, to, ...)

\method{vec_cast}{dvec.double}(x, to, ...)
}
\arguments{
\item{to}{dvec}

\item{...}{ignored}
}
\description{
Cast to dvec from double.

Cast to dvec from double.
}
\keyword{internal}
