% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_cast.R, R/vec_cast_dvec.R
\name{vec_cast.dvec.integer}
\alias{vec_cast.dvec.integer}
\title{Cast to dvec from integer}
\usage{
\method{vec_cast}{dvec.integer}(x, to, ...)

\method{vec_cast}{dvec.integer}(x, to, ...)
}
\arguments{
\item{to}{dvec}

\item{...}{ignored}
}
\description{
Cast to dvec from integer.

Cast to dvec from integer.
}
\keyword{internal}
