% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_classified.R
\name{vec_cast.factor.classified}
\alias{vec_cast.factor.classified}
\title{Cast to factor from classified}
\usage{
\method{vec_cast}{factor.classified}(x, to, ...)
}
\arguments{
\item{x}{classified}

\item{to}{factor}

\item{...}{ignored}
}
\description{
Cast to factor from classified.
}
\keyword{internal}
