% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_cast.R
\name{vec_cast.integer.dvec}
\alias{vec_cast.integer.dvec}
\title{Cast to integer from dvec}
\usage{
\method{vec_cast}{integer.dvec}(x, to, ...)
}
\arguments{
\item{to}{integer}

\item{...}{ignored}

\item{from}{dvec}
}
\description{
Cast to integer from dvec.
}
\keyword{internal}
