% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_cast.R, R/vec_cast_dvec.R
\name{vec_cast.logical.dvec}
\alias{vec_cast.logical.dvec}
\title{Cast to logical from dvec}
\usage{
\method{vec_cast}{logical.dvec}(x, to, ...)

\method{vec_cast}{logical.dvec}(x, to, ...)
}
\arguments{
\item{x}{dvec}

\item{to}{logical}

\item{...}{ignored}
}
\description{
Cast to logical from dvec.

Cast to logical from dvec.
}
\keyword{internal}
