% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_ptype2.R
\name{vec_ptype2.character.dvec}
\alias{vec_ptype2.character.dvec}
\title{Find Common Type for character, dvec}
\usage{
\method{vec_ptype2}{character.dvec}(x, y, ...)
}
\arguments{
\item{x}{character}

\item{y}{dvec}

\item{...}{ignored}
}
\description{
Find common type for character, dvec.
}
\keyword{internal}
