% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_classified.R
\name{vec_ptype2.classified.factor}
\alias{vec_ptype2.classified.factor}
\title{Find Common Type for classified, factor}
\usage{
\method{vec_ptype2}{classified.factor}(x, y, ...)
}
\arguments{
\item{x}{classified}

\item{y}{factor}

\item{...}{ignored}
}
\description{
Find common type for classified, factor.
}
\keyword{internal}
