% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_decorated.R
\name{vec_ptype2.data.frame.decorated}
\alias{vec_ptype2.data.frame.decorated}
\title{Determine Common Type for  Data.frame and decorated}
\usage{
\method{vec_ptype2}{data.frame.decorated}(x, y, ...)
}
\arguments{
\item{x}{data.frame}

\item{y}{decorated}

\item{...}{passed arguments}
}
\value{
decorated
}
\description{
Determines common type for data.frame and decorated
}
