% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_decorated.R
\name{vec_ptype2.decorated.data.frame}
\alias{vec_ptype2.decorated.data.frame}
\title{Determine Common Type for Decorated and Data.frame}
\usage{
\method{vec_ptype2}{decorated.data.frame}(x, y, ...)
}
\arguments{
\item{x}{decorated}

\item{y}{data.frame}

\item{...}{passed arguments}
}
\value{
decorated
}
\description{
Determines common type for decorated and data.frame.
}
\keyword{internal}
