% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_decorated.R
\name{vec_ptype2.decorated.decorated}
\alias{vec_ptype2.decorated.decorated}
\title{Determine Common Type for Decorated}
\usage{
\method{vec_ptype2}{decorated.decorated}(x, y, ...)
}
\arguments{
\item{x}{decorated}

\item{y}{decorated}

\item{...}{passed arguments}
}
\value{
decorated
}
\description{
Determines common type for decorated
}
