% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_ptype2.R
\name{vec_ptype2.double.dvec}
\alias{vec_ptype2.double.dvec}
\title{Find Common Type for double, dvec}
\usage{
\method{vec_ptype2}{double.dvec}(x, y, ...)
}
\arguments{
\item{x}{double}

\item{y}{dvec}

\item{...}{passed arguments}
}
\description{
Find common type for double, dvec.
}
\keyword{internal}
