% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_ptype2.R, R/vec_ptype2_dvec.R
\name{vec_ptype2.dvec.character}
\alias{vec_ptype2.dvec.character}
\title{Find Common Type for dvec, dvec}
\usage{
\method{vec_ptype2}{dvec.character}(x, y, ...)

\method{vec_ptype2}{dvec.character}(x, y, ...)
}
\arguments{
\item{x}{dvec}

\item{y}{dvec}

\item{...}{ignored}
}
\description{
Find common type for dvec, dvec.

Find common type for dvec, dvec.
}
\keyword{internal}
