% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_ptype2.R
\name{vec_ptype2.dvec.double}
\alias{vec_ptype2.dvec.double}
\title{Find Common Type for dvec, double}
\usage{
\method{vec_ptype2}{dvec.double}(x, y, ...)
}
\arguments{
\item{x}{dvec}

\item{y}{double}

\item{...}{passed arguments}
}
\description{
Find common type for dvec, double.
}
\keyword{internal}
