% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs.R
\name{vec_ptype2.dvec.dvec}
\alias{vec_ptype2.dvec.dvec}
\title{Coerce Decorated Vector to Decorated Vector}
\usage{
\method{vec_ptype2}{dvec.dvec}(x, y, ...)
}
\arguments{
\item{x}{dvec}

\item{y}{dvec}

\item{...}{passed arguments}
}
\description{
Coerces Decorated Vector to Decorated Vector.  Supports \code{\link[tidyr]{pivot_wider}}, etc.
}
\examples{
vctrs::vec_ptype2(as_dvec('bar'), as_dvec(1))
vctrs::vec_ptype2(as_dvec(1), as_dvec('foo'))
}
\keyword{internal}
