% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_ptype2.R
\name{vec_ptype2.dvec.integer}
\alias{vec_ptype2.dvec.integer}
\title{Find Common Type for dvec, integer}
\usage{
\method{vec_ptype2}{dvec.integer}(x, y, ...)
}
\arguments{
\item{x}{dvec}

\item{y}{integer}

\item{...}{ignored}
}
\description{
Find common type for dvec, integer.
}
\keyword{internal}
