% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_classified.R
\name{vec_ptype2.factor.classified}
\alias{vec_ptype2.factor.classified}
\title{Find Common Type for factor, classified}
\usage{
\method{vec_ptype2}{factor.classified}(x, y, ...)
}
\arguments{
\item{x}{factor}

\item{y}{classified}

\item{...}{ignored}
}
\description{
Find common type for factor, classified.
}
\keyword{internal}
