% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_ptype2.R
\name{vec_ptype2.integer.dvec}
\alias{vec_ptype2.integer.dvec}
\title{Find Common Type for integer, dvec}
\usage{
\method{vec_ptype2}{integer.dvec}(x, y, ...)
}
\arguments{
\item{x}{integer}

\item{y}{dvec}

\item{...}{ignored}
}
\description{
Find common type for integer, dvec.
}
\keyword{internal}
