% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{vec_ptype_abbr.dvec}
\alias{vec_ptype_abbr.dvec}
\title{Abbreviate Decorated Vector}
\usage{
\method{vec_ptype_abbr}{dvec}(x, ...)
}
\arguments{
\item{x}{classified}

\item{...}{ignored}
}
\value{
character
}
\description{
Abbreviated class name for dvec.
}
\examples{
cat(vec_ptype_abbr(as_dvec(0)))
}
\keyword{internal}
