% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-to-plotmath.R
\name{wiki_to_gg}
\alias{wiki_to_gg}
\title{Render Wiki Symbol in Plot}
\usage{
wiki_to_gg(x, ...)
}
\arguments{
\item{x}{wikisymbol; see \code{\link{as_wikisymbol}}}

\item{...}{ignored arguments}
}
\description{
For diagnostic purposes, render wikisymbol in a ggplot.
}
\examples{
wiki_to_gg('1 joule^\\\\*. ~1 kg m^2./s^2')
}
\keyword{internal}
