% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-to-latex.R
\name{wiki_to_latex}
\alias{wiki_to_latex}
\title{Convert One Wiki Symbol to Latex}
\usage{
wiki_to_latex(x, italics = FALSE, math = TRUE, ...)
}
\arguments{
\item{x}{character}

\item{italics}{whether to use italics or not (default: no)}

\item{math}{whether to wrap in math environment (default: yes)}

\item{...}{ignored}
}
\value{
character
}
\description{
Converts one wiki symbol to latex.
See description for \code{\link{as_wikisymbol}}.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_wikisymbol(x)
lapply(x, wiki_to_latex)
library(magrittr)
'joule^\\\\*. ~1 kg m^2./s^2' \%>\% wiki_to_latex
'gravitational force (kg\\\\.m/s^2.)' \%>\% wiki_to_latex
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol.symbolic_units}()},
\code{\link{as_wikisymbol.unit_string}()},
\code{\link{as_wikisymbol.units}()},
\code{\link{as_wikisymbol}()},
\code{\link{wiki_to_plotmath}()},
\code{\link{wikitoken}()}
}
\concept{wikisymbol}
\keyword{internal}
