% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-to-plotmath.R
\name{wiki_to_plotmath}
\alias{wiki_to_plotmath}
\title{Convert One Wiki Symbol to Plotmath}
\usage{
wiki_to_plotmath(x, close = character(0), ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored}
}
\value{
character
}
\description{
Converts one wiki symbol to plotmath.
See description for \code{\link{as_wikisymbol}}.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
lapply(x, wiki_to_plotmath)
library(magrittr)
wiki_to_plotmath('^*') \%>\% parse(text = .)
wiki_to_plotmath('^\\\\*') \%>\% parse(text = .)
wiki_to_plotmath('^\\\\*.') \%>\% parse(text = .)
wiki_to_plotmath('^\\\\.') \%>\% parse(text = .)
wiki_to_plotmath('^\\\\\\\\') \%>\% parse(text = .)
wiki_to_plotmath('\\\\\\\\') \%>\% parse(text = .)
wiki_to_plotmath('^\\\\^') \%>\% parse(text = .)
wiki_to_plotmath('^\\\\_') \%>\% parse(text = .)
wiki_to_plotmath('\\\\^') \%>\% parse(text = .)
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol.symbolic_units}()},
\code{\link{as_wikisymbol.unit_string}()},
\code{\link{as_wikisymbol.units}()},
\code{\link{as_wikisymbol}()},
\code{\link{wiki_to_latex}()}
}
\concept{wikisymbol}
\keyword{internal}
