% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisym2latex.R
\name{wikisym2latex_}
\alias{wikisym2latex_}
\title{Convert One Wiki Symbol to Latex}
\usage{
wikisym2latex_(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored}
}
\value{
character
}
\description{
Converts one wiki symbol to latex.
See description for \code{\link{as_wikisymbol}}.
These have special meaning in Wiki Symbol
and may be escaped with a backslash:
\code{\ * ^ _ .}.
These have special meaning in latex:
\code{& \% $ # _ { } ~ ^ \ }.
These have special meaning in both
wikisymbol and latex:
\code{\ ^ _ }.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_wikisymbol(x)
lapply(x, wikisym2latex_)
library(magrittr)
'joule^\\\\*. ~1 kg m^2./s^2' \%>\% wikisym2latex_
'gravitational force (kg\\\\.m/s^2.)' \%>\% wikisym2latex_
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol.symbolic_units}()},
\code{\link{as_wikisymbol}()},
\code{\link{wikisym2plotmath_}()}
}
\concept{wikisymbol}
\keyword{internal}
