% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{wikisym2plotmath}
\alias{wikisym2plotmath}
\title{Convert Wiki Symbol to Plotmath}
\usage{
wikisym2plotmath(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored}
}
\value{
expression
}
\description{
Converts wiki symbol to plotmath.  Vectorized version of \code{\link{wikisym2plotmath_}}.
}
\seealso{
Other formatters: 
\code{\link{wikisym2plotmath_}()}
}
\concept{formatters}
\keyword{internal}
