% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{wikisym2plotmath_}
\alias{wikisym2plotmath_}
\alias{wikisym}
\alias{wikisymbol}
\title{Convert One Wiki Symbol to Plotmath}
\usage{
wikisym2plotmath_(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored}
}
\value{
expression
}
\description{
Converts one wiki symbol to plotmath.  A Wiki symbol is simple text with arbitrarily nested subscript (\code{_}) and superscript (\code{^}) groupings.  Use dot (\code{.}) to explicitly terminate a grouping, and use backslash-dot (\code{\.}) for a literal dot.  Examples: \code{V_c./F}. Trailing dots need not be supplied. Leading/trailing whitespace is removed. Tab character not allowed.
}
\examples{
wikisym2plotmath_('V_c./F')
wikisym2plotmath_('AUC_ss')
wikisym2plotmath_('C_max_ss')
wikisym2plotmath_('var^eta_j')
}
\seealso{
Other formatters: 
\code{\link{wikisym2plotmath}()}

Other lab: 
\code{\link{as_lab.character}()},
\code{\link{as_lab.list}()},
\code{\link{as_lab}()},
\code{\link{ggplot.decorated}()},
\code{\link{print.ag}()},
\code{\link{singularity}()}
}
\concept{formatters}
\concept{lab}
