% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{alias.data.frame}
\alias{alias.data.frame}
\title{Alias a Data Frame}
\usage{
\method{alias}{data.frame}(object, ...)
}
\arguments{
\item{object}{data.frame}

\item{...}{ignored arguments}
}
\value{
aliased data.frame
}
\description{
Aliases a data.frame.
Replaces column names with labels, where present.
Stores column name as 'alias' attribute.
}
\examples{
x <- data.frame(x = 1:10, y = 1:10, z = 1:10)
attr(x$x, 'label') <- 'Independent Value'
attr(x$y, 'label') <- 'Dependent Value'
x
alias(x)
}
\seealso{
Other labels: 
\code{\link{append_units.data.frame}()},
\code{\link{append_units.default}()},
\code{\link{append_units}()}
}
\concept{labels}
