% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified.R
\name{as_classified.factor}
\alias{as_classified.factor}
\title{Coerce Factor to Classified}
\usage{
\method{as_classified}{factor}(x, ...)
}
\arguments{
\item{x}{factor}

\item{...}{ignored arguments}
}
\value{
class 'classified' 'factor'
}
\description{
Coerce factor to classified.
Creates a factor that retains attributes during subsetting.
}
\examples{
class(as_classified(factor(letters)))
}
\seealso{
Other classified: 
\code{\link{[.classified}()},
\code{\link{as_classified}()}

Other interface: 
\code{\link{conditionalize.data.frame}()},
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{ggplot.decorated}()},
\code{\link{ggready.data.frame}()},
\code{\link{io_csv.character}()},
\code{\link{io_csv.data.frame}()},
\code{\link{io_table.character}()},
\code{\link{io_table.data.frame}()},
\code{\link{io_yamlet.character}()},
\code{\link{io_yamlet.data.frame}()},
\code{\link{is_parseable.default}()},
\code{\link{read_yamlet}()},
\code{\link{resolve.data.frame}()},
\code{\link{write_yamlet}()}
}
\concept{classified}
\concept{interface}
