% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified.R
\name{as_classified.factor}
\alias{as_classified.factor}
\title{Coerce Factor to Classified}
\usage{
\method{as_classified}{factor}(x, ...)
}
\arguments{
\item{x}{factor}

\item{...}{ignored arguments}
}
\value{
class 'classified' 'factor'
}
\description{
Coerce factor to classified.
Creates a factor that retains attributes during subsetting.
}
\examples{
class(as_classified(factor(letters)))
}
\seealso{
Other classified: 
\code{\link{[.classified}()},
\code{\link{as_classified}()}
}
\concept{classified}
\keyword{internal}
