% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R, R/dplyr.R
\name{as_decorated}
\alias{as_decorated}
\title{Coerce to Decorated}
\usage{
as_decorated(x, ...)

as_decorated(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\value{
decorated
}
\description{
Coerces to class 'decorated'. Generic, with default method.

Coerces to decorated by subclassing.
}
\examples{
class(Puromycin)
class(as_decorated(Puromycin))
}
\seealso{
Other decorate: 
\code{\link{as_decorated.default}()},
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{decorate.list}()},
\code{\link{decorate}()},
\code{\link{decorations.data.frame}()},
\code{\link{decorations}()},
\code{\link{print.decorations}()},
\code{\link{redecorate}()}

Other dplyr: 
\code{\link{anti_join.decorated}()},
\code{\link{arrange.decorated}()},
\code{\link{filter.decorated}()},
\code{\link{full_join.decorated}()},
\code{\link{group_by.decorated}()},
\code{\link{inner_join.decorated}()},
\code{\link{left_join.decorated}()},
\code{\link{mutate.decorated}()},
\code{\link{right_join.decorated}()},
\code{\link{select.decorated}()},
\code{\link{semi_join.decorated}()},
\code{\link{slice.decorated}()},
\code{\link{summarise.decorated}()},
\code{\link{summarize.decorated}()}
}
\concept{decorate}
\concept{dplyr}
\keyword{internal}
