% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{as_decorated}
\alias{as_decorated}
\title{Coerce to Decorated}
\usage{
as_decorated(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
decorated
}
\description{
Coerces to class 'decorated'. Generic, with method \code{\link{as_decorated.default}}.
}
\examples{
class(Puromycin)
class(as_decorated(Puromycin))
}
\seealso{
Other decorate: 
\code{\link{as_decorated.default}()},
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{decorate.list}()},
\code{\link{decorate}()},
\code{\link{decorations.data.frame}()},
\code{\link{decorations}()},
\code{\link{redecorate}()}
}
\concept{decorate}
\keyword{internal}
