% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{as_dvec.logical}
\alias{as_dvec.logical}
\title{Coerce Logical to Decorated Vector}
\usage{
\method{as_dvec}{logical}(x, ...)
}
\arguments{
\item{x}{logical}

\item{...}{ignored arguments}
}
\value{
dvec
}
\description{
Coerces logical to decorated Vector. Expects values TRUE and/or FALSE,
and coerces these to character.  Assigns class 'dvec'.
}
\examples{
as_dvec(FALSE:TRUE)
}
\seealso{
Other dvec: 
\code{\link{[.dvec}()},
\code{\link{[<-.dvec}()},
\code{\link{[[.dvec}()},
\code{\link{[[<-.dvec}()},
\code{\link{as_dvec.character}()},
\code{\link{as_dvec.integer}()},
\code{\link{as_dvec.numeric}()},
\code{\link{as_dvec}()},
\code{\link{c.dvec}()},
\code{\link{format.dvec}()},
\code{\link{length.dvec}()},
\code{\link{print.dvec}()}
}
\concept{dvec}
