% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{as_dvec.numeric}
\alias{as_dvec.numeric}
\title{Coerce Numeric to Decorated Vector}
\usage{
\method{as_dvec}{numeric}(x, ...)
}
\arguments{
\item{x}{numeric}

\item{...}{attributes to assign}
}
\value{
dvec
}
\description{
Coerces numeric to decorated vector.  Assigns class 'dvec' and any named attributes in dots.
}
\examples{
as_dvec(c(10.3, 1.2))
as_dvec(1, label = 'yin')
as_dvec(structure(1, label = 'yin'))
as_dvec(structure(1, label = 'yin'), label = 'yang')

}
\seealso{
Other dvec: 
\code{\link{[.dvec}()},
\code{\link{[<-.dvec}()},
\code{\link{[[.dvec}()},
\code{\link{[[<-.dvec}()},
\code{\link{as.data.frame.dvec}()},
\code{\link{as_dvec.character}()},
\code{\link{as_dvec.complex}()},
\code{\link{as_dvec.dvec}()},
\code{\link{as_dvec.integer}()},
\code{\link{as_dvec.logical}()},
\code{\link{as_dvec}()},
\code{\link{c.dvec}()},
\code{\link{classified.dvec}()},
\code{\link{desolve.dvec}()},
\code{\link{explicit_guide.data.frame}()},
\code{\link{format.dvec}()},
\code{\link{implicit_guide.dvec}()},
\code{\link{length.dvec}()},
\code{\link{print.dvec}()},
\code{\link{resolve.classified}()},
\code{\link{resolve.dvec}()},
\code{\link{unclassified.dvec}()}
}
\concept{dvec}
\keyword{internal}
