% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{as_dvec}
\alias{as_dvec}
\title{Coerce to Decorated Vector}
\usage{
as_dvec(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{ignored arguments}
}
\value{
dvec
}
\description{
Coerces to Decorated Vector. Generic, with methods
\code{\link{as_dvec.logical}},
\code{\link{as_dvec.integer}},
\code{\link{as_dvec.numeric}},
\code{\link{as_dvec.complex}}, and
\code{\link{as_dvec.character}}.
}
\examples{
as_dvec(0)
}
\seealso{
Other dvec: 
\code{\link{[.dvec}()},
\code{\link{[<-.dvec}()},
\code{\link{[[.dvec}()},
\code{\link{[[<-.dvec}()},
\code{\link{as.data.frame.dvec}()},
\code{\link{as_dvec.character}()},
\code{\link{as_dvec.complex}()},
\code{\link{as_dvec.dvec}()},
\code{\link{as_dvec.integer}()},
\code{\link{as_dvec.logical}()},
\code{\link{as_dvec.numeric}()},
\code{\link{c.dvec}()},
\code{\link{classified.dvec}()},
\code{\link{desolve.dvec}()},
\code{\link{explicit_guide.data.frame}()},
\code{\link{format.dvec}()},
\code{\link{implicit_guide.dvec}()},
\code{\link{length.dvec}()},
\code{\link{print.dvec}()},
\code{\link{resolve.classified}()},
\code{\link{resolve.dvec}()},
\code{\link{unclassified.dvec}()}
}
\concept{dvec}
\keyword{internal}
