% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agplot.R
\name{as_lab}
\alias{as_lab}
\title{Coerce to Axis Label}
\usage{
as_lab(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods; typically length-one character
}
\description{
Converts to axis label. Generic, with method \code{\link{as_lab.list}}.
}
\seealso{
Other lab: 
\code{\link{as_lab.list}()},
\code{\link{ggplot.decorated}()},
\code{\link{print.ag}()},
\code{\link{singularity}()}
}
\concept{lab}
\keyword{internal}
