% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{as_latex}
\alias{as_latex}
\title{Coerce to Latex}
\usage{
as_latex(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
latex
}
\description{
Coerce to latex.  Generic, with method
\code{\link{as_latex.wikisymbol}}.
}
\examples{
example(as_latex.wikisymbol)
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol.symbolic_units}()},
\code{\link{as_wikisymbol.unit_string}()},
\code{\link{as_wikisymbol.units}()},
\code{\link{as_wikisymbol}()},
\code{\link{wiki_to_latex}()},
\code{\link{wiki_to_plotmath}()}
}
\concept{wikisymbol}
\keyword{internal}
