% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-to-latex.R
\name{as_preview}
\alias{as_preview}
\title{Preview Something}
\usage{
as_preview(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Creates a preview.
Generic, with methods for latex and plotmath.
}
\examples{
library(magrittr)
\dontrun{
'V_c./F' \%>\% as_wikisymbol \%>\% as_latex \%>\% as_preview
}
'V_c./F' \%>\% as_wikisymbol \%>\% as_plotmath \%>\% as_preview
}
\seealso{
Other preview: 
\code{\link{as_preview.latex}()},
\code{\link{as_preview.plotmath}()},
\code{\link{ggplot.plotmath}()},
\code{\link{ggplot.wikisymbol}()}
}
\concept{preview}
