% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{as_wikisymbol}
\alias{as_wikisymbol}
\title{Coerce to Wiki Symbol}
\usage{
as_wikisymbol(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
wikisymbol
}
\description{
Coerces to class 'wikisymbol'. Generic,
with method \code{\link{as_wikisymbol.character}}.
A Wiki symbol is simple text with arbitrarily
nested subscript (\code{_}) and superscript
(\code{^}) groupings.  Use dot (\code{.})
to explicitly terminate a grouping.
Use asterisk(\code{*}) to suggest multiplication.
Escape specials with a backslash.
Trailing dots need not be supplied.
Leading/trailing whitespace is removed.
Tab character not allowed.
}
\examples{
example(as_wikisymbol.character)
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol.symbolic_units}()},
\code{\link{wikisym2latex_}()},
\code{\link{wikisym2plotmath_}()}
}
\concept{wikisymbol}
\keyword{internal}
