% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{as_wikisymbol}
\alias{as_wikisymbol}
\title{Coerce to Wiki Symbol}
\usage{
as_wikisymbol(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
wikisymbol
}
\description{
Coerces to class 'wikisymbol'. Generic,
with method \code{\link{as_wikisymbol.character}}.
A wiki symbol is simple text expressing
arbitrarily nested subscripts (\code{x_y_z})
and superscripts (\code{x^y^z}). A dot
(\code{x^y._z}) explicitly terminates
a group. An asterisk (\code{*}) suggests
multiplication. Special characters
may be escaped with a backslash.
Convert to plotmath with\code{\link{as_plotmath}}
and to latex with \code{\link{as_latex}}.
}
\examples{
library(magrittr)
library(dplyr)
library(ggplot2)
library(latexpdf)

'j^*' \%>\% as_wikisymbol \%>\% as_plotmath \%>\% as.character
as_wikisymbol \%>\%
as_plotmath #\%>\%
as_expression
data.frame(y=1:10, x=1:10) \%>\%
decorate("x: Omega joule^\\\\*. ~1 kg*m^2./s^2. [\%]") \%>\%
mutate(
 x = structure(x, label = x \%>\% attr('label') \%>\%
 as_wikisymbol \%>\%
 as_plotmath \%>\%
 as.expression)
) \%>\%
 ggplot(aes(x, y))

data.frame(y=1:10, x=1:10) \%>\%
decorate("x: gravitational force - gamma (kg\\\\.m/s^2.)") \%>\%
mutate(x = structure(x, label = x \%>\% attr('label') \%>\%
as_wikisymbol \%>\%
as_plotmath \%>\%
as.expression)) \%>\%
ggplot(aes(x, y))

path <- tempfile()
data.frame(
 stringsAsFactors = FALSE,
 wikisymbol = c(
   'Omega joule^\\\\*. ~1 kg*m^2./s^2. [\%]',
   'gravitational force - gamma (kg\\\\.m/s^2.)'
 )
) \%>\%
 transmute(
   latex = wikisymbol \%>\%
   as_wikisymbol \%>\%
   as_latex
 ) \%>\%
 as.pdf(reserve = FALSE, wider = -70)
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol.symbolic_units}()},
\code{\link{as_wikisymbol.unit_string}()},
\code{\link{as_wikisymbol.units}()},
\code{\link{wiki_to_latex}()},
\code{\link{wiki_to_plotmath}()},
\code{\link{wikitoken}()}
}
\concept{wikisymbol}
\keyword{internal}
