% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{as_wikisymbol.symbolic_units}
\alias{as_wikisymbol.symbolic_units}
\title{Coerce Symbolic Units to Wiki Symbol.}
\usage{
\method{as_wikisymbol}{symbolic_units}(x, ...)
}
\arguments{
\item{x}{units; see \code{\link[units]{as_units}}}

\item{...}{ignored arguments}
}
\value{
units
}
\description{
Coerces symbolic units to wikisymbol.  A literal dot
means different things in wikisymbol vs. units,
and there may be some other subtleties as well.
}
\examples{
library(units)
x <- as_units('kg.m/s^2')
names(attributes(x))
y <- attr(x,'units')
class(y)
as.character(y)
as.character(attr(x, 'units'))
as_wikisymbol(y)
library(magrittr)
'kg.m^2/s^2' \%>\% as_units \%>\% attr('units') \%>\% as_wikisymbol
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol}()},
\code{\link{wikisym2latex_}()},
\code{\link{wikisym2plotmath_}()}
}
\concept{wikisymbol}
