% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{as_yam}
\alias{as_yam}
\title{Coerce to Yam}
\usage{
as_yam(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
list
}
\description{
Coerce to yam, a precursor to yamlet.  Generic, with character
method: \code{\link{as_yam.character}}.
}
\seealso{
Other yam: 
\code{\link{as.character.yam}()},
\code{\link{as_yam.character}()},
\code{\link{as_yam.yamlet}()}
}
\concept{yam}
\keyword{internal}
