% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{as_yamlet.character}
\alias{as_yamlet.character}
\title{Convert Character To yamlet Format}
\usage{
\method{as_yamlet}{character}(x, default_keys = list("label", "guide"), ...)
}
\arguments{
\item{x}{length-one filepath or actual data}

\item{default_keys}{character: default keys for the first n anonymous members of each element}

\item{...}{passed arguments}
}
\value{
yamlet: a named list with default keys applied
}
\description{
Converts character to yamlet format.
Length-one character is understood as a file path
if the file exists.  Otherwise, it is treated as data.
The file is a mapping of (nested) sequences,
where map keys are data item names, and
sequences represent data item attributes.
Attributes may be named or anonymous.
A map key '_keys' identifies a sequence of
key names that over-ride \code{default_keys}.
}
\examples{
as_yamlet(system.file(package = 'yamlet', 'extdata','yamlet.yaml'))
as_yamlet('ID: subject identifier')
as_yamlet(c('id: subject','amt: dose'))
as_yamlet(c('id: subject\namt: dose'))

}
\seealso{
Other yamlet: 
\code{\link{as_yamlet.yam}()},
\code{\link{as_yamlet}()}
}
\concept{yamlet}
