% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{as_yamlet.data.frame}
\alias{as_yamlet.data.frame}
\title{Coerce Data Frame to Yamlet}
\usage{
\method{as_yamlet}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{passed to \code{\link{decorations}}}
}
\value{
yamlet
}
\description{
Coerces data.frame to yamlet by calling {\code{\link{decorations.data.frame}}}.
}
\examples{
library(csv)
file <- system.file(package = 'yamlet', 'extdata','quinidine.csv')
x <- decorate(as.csv(file))
as_yamlet(x)
}
\seealso{
Other yamlet: 
\code{\link{[.yamlet}()},
\code{\link{as.character.yamlet}()},
\code{\link{as_yamlet.character}()},
\code{\link{as_yamlet.list}()},
\code{\link{as_yamlet.yamlet}()},
\code{\link{as_yamlet.yam}()},
\code{\link{as_yamlet}()},
\code{\link{print.yamlet}()}
}
\concept{yamlet}
\keyword{internal}
