% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{as_yamlet}
\alias{as_yamlet}
\title{Coerce to Yamlet}
\usage{
as_yamlet(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
a named list
}
\description{
Coerces something to yamlet format. If the object
or user specifies default keys, these are applied.
See \code{\link{as_yamlet.character}}.
}
\examples{
file <- system.file(package = 'yamlet', 'extdata','quinidine.yaml')
file
identical(as_yamlet(as_yam(file)), as_yamlet(file))

# Read yamlet from storage and apply default keys.
as_yamlet(file)

}
\seealso{
Other yamlet: 
\code{\link{[.yamlet}()},
\code{\link{as.character.yamlet}()},
\code{\link{as_yamlet.character}()},
\code{\link{as_yamlet.data.frame}()},
\code{\link{as_yamlet.list}()},
\code{\link{as_yamlet.yamlet}()},
\code{\link{as_yamlet.yam}()},
\code{\link{print.yamlet}()}
}
\concept{yamlet}
\keyword{internal}
