% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{decorate.character}
\alias{decorate.character}
\title{Decorate Character}
\usage{
\method{decorate}{character}(x, meta = NULL, ...)
}
\arguments{
\item{x}{file path for csv}

\item{meta}{file path for corresponding yaml metadata, or a yamlet}

\item{...}{passed arguments}
}
\description{
Treats \code{x} as a csv file path. By default,
metadata is sought from a file with the same
base but the 'yaml' extension.
}
\examples{
file <- system.file(package = 'yamlet', 'extdata','yamlet.csv')
meta <- system.file(package = 'yamlet', 'extdata','yamlet.yaml')
x <- decorate(file)
sapply(x, attr, 'label')

}
\seealso{
Other decorate: \code{\link{decorate.data.frame}},
  \code{\link{decorate.list}}, \code{\link{decorate}}
}
\concept{decorate}
