% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate}
\alias{decorate_generic}
\title{Decorate a List-like Object}
\usage{
decorate(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
a list-like object, typically data.frame
}
\description{
Decorates a list-like object. Generic.
See \code{\link{decorate.character}}.
}
\examples{
library(csv)
file <- system.file(package = 'yamlet', 'extdata','quinidine.csv')
x <- decorate(as.csv(file))
identical(decorate(as.csv(file)), decorate(file))
decorations(x)


}
\seealso{
Other decorate: 
\code{\link{as_decorated.default}()},
\code{\link{as_decorated}()},
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{decorate.list}()},
\code{\link{decorate_groups.data.frame}()},
\code{\link{decorate_groups_generic}},
\code{\link{decorations.data.frame}()},
\code{\link{decorations_generic}},
\code{\link{decorations_groups.data.frame}()},
\code{\link{decorations_groups_generic}},
\code{\link{group_by_decorations.data.frame}()},
\code{\link{group_by_decorations_generic}},
\code{\link{redecorate}()}
}
\concept{decorate}
\keyword{internal}
