% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{decorations.data.frame}
\alias{decorations.data.frame}
\title{Retrieve Decorations for Data.frame}
\usage{
\method{decorations}{data.frame}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
named list
}
\description{
Retrieve the decorations of a data.frame; i.e., the metadata
used to decorate it. Returns a list with same names as the data.frame.
}
\examples{
library(csv)
file <- system.file(package = 'yamlet', 'extdata','yamlet.csv')
x <- decorate(as.csv(file))
decorations(x)
}
\seealso{
Other decorate: 
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{decorate.list}()},
\code{\link{decorate}()},
\code{\link{decorations}()}
}
\concept{decorate}
