% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{decorations}
\alias{decorations}
\title{Retrieve Decorations}
\usage{
decorations(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Retrieve the decorations of something.
Generic, with method for data.frame.
}
\examples{
library(csv)
file <- system.file(package = 'yamlet', 'extdata','quinidine.csv')
x <- decorate(as.csv(file))
decorations(x) # prefer as_yamlet(x)
}
\seealso{
Other decorate: 
\code{\link{decorate.character}()},
\code{\link{decorate.data.frame}()},
\code{\link{decorate.list}()},
\code{\link{decorate}()},
\code{\link{decorations.data.frame}()}
}
\concept{decorate}
\keyword{internal}
