% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desolve.R
\name{desolve}
\alias{desolve}
\title{Desolve Guide}
\usage{
desolve(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Un-resolves explicit versions of 'guide' to implicit usage.
Generic, with method \code{\link{desolve.decorated}}.
}
\examples{
example(resolve.decorated)
}
\seealso{
Other resolve: 
\code{\link{desolve.decorated}()},
\code{\link{resolve.decorated}()},
\code{\link{resolve}()}
}
\concept{resolve}
\keyword{internal}
