% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{get_labs}
\alias{get_labs}
\title{Get Labels}
\usage{
get_labs(plot)
}
\arguments{
\item{plot}{the ggplot}
}
\description{
Gets labels for a ggplot object.  Not exported, to avoid confusion.
Development version of ggplot2 implements new get_labs() interface.
This function is an abstraction that supports new vs old approaches,
solely for yamlet's interests.
See https://github.com/tidyverse/ggplot2/pull/6078.
}
