% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_ggready.R
\name{ggplot.ggready}
\alias{ggplot.ggready}
\title{Create a New ggplot for a GGready Data Frame}
\usage{
\method{ggplot}{ggready}(data, ...)
}
\arguments{
\item{data}{data.frame or similar}

\item{...}{passed to \code{\link[ggplot2]{ggplot}}}
}
\value{
return value like \code{\link[ggplot2]{ggplot}}
}
\description{
Creates a new ggplot object for a ggready data.frame.
This is the ggplot() method for class 'ggready';
it tries to implement automatic labels and units in axes and legends
in association with \code{\link{print.ggready_ggplot}}.
This approach is deprecated in favor of \code{\link{ggplot.decorated}}.
}
\examples{
example(ggready)
}
\seealso{
Other ggready: 
\code{\link{ggplot_build.ggready_ggplot}()},
\code{\link{ggready.data.frame}()},
\code{\link{ggready.decorated}()},
\code{\link{ggready}()},
\code{\link{print.ggready_ggplot}()}
}
\concept{ggready}
\keyword{internal}
