% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot_build.decorated_ggplot}
\alias{ggplot_build.decorated_ggplot}
\title{Enable Automatic Labels and Units for ggplot}
\usage{
ggplot_build.decorated_ggplot(
  x,
  ...,
  search = getOption("resolved_ggplot_search", c("expression", "title", "label"))
)
}
\arguments{
\item{x}{class 'resolved_ggplot' from \code{\link{ggplot.ggready}}}

\item{...}{passed arguments}
}
\value{
see \code{\link[ggplot2]{ggplot_build}}
}
\description{
Enable automatic labels and units for ggplot.
Substitutes column label, if present, for default.
Supports arrangements of ggplot objects.
}
\concept{resolved_ggplot}
\keyword{internal}
