% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_build.R
\name{ggplot_build.decorated_ggplot}
\alias{ggplot_build.decorated_ggplot}
\title{Enable Automatic Labels and Units for ggplot}
\usage{
\method{ggplot_build}{decorated_ggplot}(plot, ...)
}
\arguments{
\item{plot}{class 'decorated_ggplot' from \code{\link{ggplot.decorated}}}
}
\value{
see \code{\link[ggplot2]{ggplot_build}}
}
\description{
Enable automatic labels and units for ggplot.
Substitutes column label, if present, for default.
Supports arrangements of ggplot objects.
Defined similarly to \code{\link{print.decorated_ggplot}}
and respects global options 
yamlet_decorated_ggplot_search,
yamlet_decorated_ggplot_discrete, and
yamlet_decorated_ggplot_drop.
}
\seealso{
Other decorated_ggplot: 
\code{\link{ggplot.decorated}()},
\code{\link{print.decorated_ggplot}()}
}
\concept{decorated_ggplot}
\keyword{internal}
