% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot_build.decorated_ggplot}
\alias{ggplot_build.decorated_ggplot}
\title{Enable Automatic Labels and Units for ggplot}
\usage{
\method{ggplot_build}{decorated_ggplot}(
  x,
  ...,
  search = getOption("yamlet_decorated_ggplot_search", c("expression", "title", "label")),
  discrete = getOption("yamlet_decorated_ggplot_discrete", c("colour", "fill", "size",
    "shape", "linetype", "alpha")),
  drop = getOption("yamlet_decorated_ggplot_drop", TRUE)
)
}
\arguments{
\item{x}{class 'decorated_ggplot' from \code{\link{ggplot.decorated}}}

\item{...}{passed arguments}
}
\value{
see \code{\link[ggplot2]{ggplot_build}}
}
\description{
Enable automatic labels and units for ggplot.
Substitutes column label, if present, for default.
Supports arrangements of ggplot objects.
}
\seealso{
Other decorated_ggplot: 
\code{\link{ggplot.decorated}()},
\code{\link{print.decorated_ggplot}()}
}
\concept{decorated_ggplot}
\keyword{internal}
