% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agplot.R
\name{ggplot_build.dg}
\alias{ggplot_build.dg}
\title{Enable Automatic Labels and Units for ggplot}
\usage{
ggplot_build.dg(x, ...)
}
\arguments{
\item{x}{class 'dg' from \code{\link{ggplot.decorated}}}

\item{...}{passed arguments}
}
\value{
see \code{\link[ggplot2]{ggplot_build}}
}
\description{
Enable automatic labels and units for ggplot.
Substitutes column label, if present, for default.
Supports arrangements of ggplot objects.
}
\seealso{
Other dg: 
\code{\link{ggplot.decorated}()},
\code{\link{print.dg}()}
}
\concept{dg}
