% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggready.R
\name{ggready.data.frame}
\alias{ggready.data.frame}
\title{Prepare Data Frame for GGplot}
\usage{
\method{ggready}{data.frame}(x, ..., parse = getOption("yamlet_ggready_parse", TRUE))
}
\arguments{
\item{x}{object}

\item{...}{passed to \code{\link{append_units}}; may include unquoted column names}

\item{parse}{passed to \code{\link{append_units}}}
}
\value{
ggready decorated
}
\description{
Prepares data.frame for ggplot. Appends
units to label using \code{\link{append_units}}
(passing \code{style = 'plotmath'} if \code{parse}
is true, else \code{style = 'plain'}).
Enforces classes 'decorated' and 'ggready'.
}
\examples{
example(ggready)
}
\seealso{
Other ggready: 
\code{\link{ggready.decorated}()},
\code{\link{ggready}()}
}
\concept{ggready}
\keyword{internal}
