% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_res.R
\name{io_res}
\alias{io_res}
\title{Import Resolved Tables}
\usage{
io_res(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
See methods.
}
\description{
Imports 'resolved' tables as comma-separated variable.
Generic, with character method that extends \code{\link{io_csv}}.
}
\examples{
example(io_res.character)
}
\seealso{
Other io: 
\code{\link{io_csv.character}()},
\code{\link{io_csv.data.frame}()},
\code{\link{io_csv}()},
\code{\link{io_res.character}()},
\code{\link{io_table.character}()},
\code{\link{io_table.data.frame}()},
\code{\link{io_table}()},
\code{\link{io_yamlet.character}()},
\code{\link{io_yamlet.data.frame}()},
\code{\link{io_yamlet.yamlet}()},
\code{\link{io_yamlet}()}
}
\concept{io}
