% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agplot.R
\name{isConditional.default}
\alias{isConditional.default}
\title{Test Object is Conditional by Default}
\usage{
\method{isConditional}{default}(x, ...)
}
\arguments{
\item{x}{default}

\item{...}{passed arguments}
}
\value{
logical
}
\description{
Tests whether object is conditional by default. Coerces to list.
}
\seealso{
Other conditional: 
\code{\link{isConditional.list}()},
\code{\link{isConditional}()}
}
\concept{conditional}
\keyword{internal}
