% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional.R
\name{isConditional.list}
\alias{isConditional.list}
\title{Test List is Conditional}
\usage{
\method{isConditional}{list}(x, data, ...)
}
\arguments{
\item{x}{list}

\item{data}{environment for variable lookup}

\item{...}{passed arguments}
}
\value{
length-one logical
}
\description{
Tests whether a list is conditional.
Evaluates names of x on data and looks for meaningful result.
Returns TRUE if list has names and
all evaluate to logicals with length equal
to number of rows in data.
}
\seealso{
Other conditional: 
\code{\link{isConditional.default}()},
\code{\link{isConditional}()}
}
\concept{conditional}
\keyword{internal}
