% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agplot.R
\name{isConditional}
\alias{isConditional}
\title{Test Object is Conditional}
\usage{
isConditional(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{passed arguments}
}
\value{
logical
}
\description{
Tests whether object is conditional.
}
\seealso{
Other conditional: 
\code{\link{isConditional.default}()},
\code{\link{isConditional.list}()}
}
\concept{conditional}
\keyword{internal}
