% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{print.decorated_ggplot}
\alias{print.decorated_ggplot}
\title{Substitute Expressions, Titles and Labels in ggplots}
\usage{
\method{print}{decorated_ggplot}(
  x,
  ...,
  search = getOption("decorated_ggplot_search", c("expression", "title", "label"))
)
}
\arguments{
\item{x}{class 'decorated_ggplot' from \code{\link{ggplot.decorated}}}

\item{...}{passed arguments}

\item{search}{attribute names from which to seek label substitutes}
}
\value{
see \code{\link[ggplot2]{print.ggplot}}
}
\description{
At time of printing, default labels will be used as
column names to search \code{data} for more meaningful
labels, taking first available from attributes
with names in \code{search}.
}
\examples{
example(ggplot.decorated)
}
\seealso{
Other decorated_ggplot: 
\code{\link{ggplot.decorated}()},
\code{\link{ggplot_build.decorated_ggplot}()}
}
\concept{decorated_ggplot}
