% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{print.yamlet}
\alias{print.yamlet}
\title{Print a Yamlet}
\usage{
\method{print}{yamlet}(x, ...)
}
\arguments{
\item{x}{yamlet}

\item{...}{ignored}
}
\value{
invisible(x)
}
\description{
Prints a yamlet object for interactive inspection.
}
\examples{
as_yamlet('mpg: [efficiency, mi/gallon]\nvs: [Engine, [V-shaped: 0, straight: 1]]')
}
\seealso{
Other yamlet: 
\code{\link{as_yamlet.character}()},
\code{\link{as_yamlet.yam}()},
\code{\link{as_yamlet}()}
}
\concept{yamlet}
