% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvec.R
\name{reconcile.list}
\alias{reconcile.list}
\title{Reconcile Atttributes of List Members}
\usage{
\method{reconcile}{list}(x, ...)
}
\arguments{
\item{x}{list}

\item{...}{passed arguments}
}
\value{
list (of attributes)
}
\description{
Reconciles attributes of list members. Recursively arbitrates
list members pairwise, returning the accumulated attributes.
}
\examples{
library(magrittr)
library(dplyr)
a <- data.frame(study = 1) \%>\% decorate('study: [Study, [A: 1]]')
b <- data.frame(study = 2) \%>\% decorate('study: [Study, [B: 2]]')
bind_rows(a, b) \%>\% decorations
c(a$study, b$study)
reconcile(list(a$study, b$study))
}
\keyword{internal}
