% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_units.R, R/classified.R, R/dplyr.R,
%   R/dvec.R, R/ggready.R, R/isometric.R, R/yamlet.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{latexToken}
\alias{type_sum}
\alias{filter}
\alias{as_units}
\alias{vec_ptype_abbr}
\alias{plotmathToken}
\alias{ggplot_add}
\alias{encode}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}}

  \item{encode}{\code{\link[encode]{encode}}}

  \item{ggplot2}{\code{\link[ggplot2:update_ggplot]{ggplot_add}}}

  \item{pillar}{\code{\link[pillar]{type_sum}}}

  \item{spork}{\code{\link[spork]{latexToken}}, \code{\link[spork]{plotmathToken}}}

  \item{units}{\code{\link[units:units]{as_units}}}

  \item{vctrs}{\code{\link[vctrs:vec_ptype_full]{vec_ptype_abbr}}}
}}

