% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{resolve}
\alias{resolve_generic}
\title{Resolve Guide}
\usage{
resolve(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Resolves implicit usage of default key 'guide' to explicit usage.
Generic, with methods 
\code{\link{resolve.decorated}} and 
\code{\link{resolve.dvec}}.
}
\examples{
example(resolve.decorated)
}
\seealso{
Other resolve: 
\code{\link{desolve.classified}()},
\code{\link{desolve.decorated}()},
\code{\link{desolve.dvec}()},
\code{\link{desolve_generic}},
\code{\link{resolve.classified}()},
\code{\link{resolve.decorated}()},
\code{\link{resolve.dvec}()}
}
\concept{resolve}
\keyword{internal}
