% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{resolve}
\alias{resolve}
\title{Resolve Guide}
\usage{
resolve(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Resolves implicit usage of default key 'guide' to explicit usage.
Generic, with method for data.frame.
}
\examples{
example(resolve.data.frame)
}
\seealso{
Other resolve: 
\code{\link{ggready.data.frame}()},
\code{\link{ggready}()},
\code{\link{resolve.data.frame}()}
}
\concept{resolve}
\keyword{internal}
