% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripted.R
\name{scripted}
\alias{scripted}
\title{Render Scripted Attributes of Indicated Components}
\usage{
scripted(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Renders the scripted attributes of indicated components.
Generic, with method \code{\link{scripted.default}}.
}
\examples{
example(scripted.default)
}
\seealso{
modify.default

as_spork

Other scripted: 
\code{\link{scripted.default}()}
}
\concept{scripted}
\keyword{internal}
