% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{slice.decorated}
\alias{slice.decorated}
\title{Slice Decorated}
\usage{
\method{slice}{decorated}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{...}{For \code{slice()}: <\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Integer row
values.

Provide either positive values to keep, or negative values to drop.
The values provided must be either all positive or all negative.
Indices beyond the number of rows in the input are silently ignored.

For \code{slice_helpers()}, these arguments are passed on to methods.}

\item{.preserve}{Relevant when the \code{.data} input is grouped.
If \code{.preserve = FALSE} (the default), the grouping structure
is recalculated based on the resulting data, otherwise the grouping is kept as is.}
}
\description{
Preserves class when slicing decorated.
}
\seealso{
Other dplyr: 
\code{\link{anti_join.decorated}()},
\code{\link{arrange.decorated}()},
\code{\link{full_join.decorated}()},
\code{\link{group_by.decorated}()},
\code{\link{inner_join.decorated}()},
\code{\link{left_join.decorated}()},
\code{\link{mutate.decorated}()},
\code{\link{right_join.decorated}()},
\code{\link{select.decorated}()},
\code{\link{semi_join.decorated}()},
\code{\link{summarise.decorated}()},
\code{\link{summarize.decorated}()}
}
\concept{dplyr}
\keyword{internal}
