% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{[.yamlet}
\alias{[.yamlet}
\title{Subset Yamlet}
\usage{
\method{[}{yamlet}(x, ...)
}
\arguments{
\item{x}{object to subset}

\item{...}{passed to next method}
}
\value{
yamlet
}
\description{
Subsets yamlet. Preserves class, since a subset of yamlet is still yamlet.
}
\examples{
meta <- system.file(package = 'yamlet', 'extdata','quinidine.yaml')
meta <- as_yamlet(meta)
class(meta)
stopifnot(inherits(meta[1:2],'yamlet'))
}
\seealso{
Other yamlet: 
\code{\link{as.character.yamlet}()},
\code{\link{as_yamlet.character}()},
\code{\link{as_yamlet.data.frame}()},
\code{\link{as_yamlet.list}()},
\code{\link{as_yamlet.yamlet}()},
\code{\link{as_yamlet.yam}()},
\code{\link{as_yamlet}()},
\code{\link{print.yamlet}()}
}
\concept{yamlet}
\keyword{internal}
