% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclassified.R
\name{unclassified.classified}
\alias{unclassified.classified}
\title{Unclassify Classified}
\usage{
\method{unclassified}{classified}(x, ..., persistence = getOption("yamlet_persistence", TRUE))
}
\arguments{
\item{x}{classified}

\item{...}{ignored}

\item{persistence}{whether to reclass as dvec}
}
\value{
vector
}
\description{
Unclassifies classified.  Uses codelist attribute
to restore original values, preserving other attributes
(and rebuilding codelist).
}
\examples{
example(unclassified.data.frame)
}
\seealso{
Other classified: 
\code{\link{[.classified}()},
\code{\link{[<-.classified}()},
\code{\link{[[.classified}()},
\code{\link{[[<-.classified}()},
\code{\link{as.integer.classified}()},
\code{\link{c.classified}()},
\code{\link{classified}()},
\code{\link{classified.classified}()},
\code{\link{classified.data.frame}()},
\code{\link{classified.default}()},
\code{\link{classified.dvec}()},
\code{\link{classified.factor}()},
\code{\link{desolve.classified}()},
\code{\link{unclassified}()},
\code{\link{unclassified.data.frame}()}
}
\concept{classified}
\keyword{internal}
