% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_cast.R
\name{vec_cast.dvec.dvec}
\alias{vec_cast.dvec.dvec}
\title{Cast to dvec from dvec}
\usage{
\method{vec_cast}{dvec.dvec}(x, to, ...)
}
\arguments{
\item{x}{dvec}

\item{to}{dvec}

\item{...}{passed arguments}
}
\description{
Cast to dvec from dvec.
}
\keyword{internal}
