% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_cast.R
\name{vec_cast.dvec.logical}
\alias{vec_cast.dvec.logical}
\title{Cast to dvec from logical}
\usage{
\method{vec_cast}{dvec.logical}(x, to, ...)
}
\arguments{
\item{x}{logical}

\item{to}{dvec}

\item{...}{passed arguments}
}
\description{
Cast to dvec from logical
}
\keyword{internal}
