% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_ptype2.R
\name{vec_ptype2.complex.dvec}
\alias{vec_ptype2.complex.dvec}
\title{Find Common Type for complex, dvec}
\usage{
\method{vec_ptype2}{complex.dvec}(x, y, ...)
}
\arguments{
\item{x}{complex}

\item{y}{dvec}

\item{...}{passed arguments}
}
\description{
Find common type for complex, dvec.
}
\keyword{internal}
