% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_ptype2.R
\name{vec_ptype2.dvec.complex}
\alias{vec_ptype2.dvec.complex}
\title{Find Common Type for dvec, complex}
\usage{
\method{vec_ptype2}{dvec.complex}(x, y, ...)
}
\arguments{
\item{x}{dvec}

\item{y}{complex}

\item{...}{passed arguments}
}
\description{
Find common type for dvec, complex.
}
\keyword{internal}
