% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_ptype2.R
\name{vec_ptype2.dvec.double}
\alias{vec_ptype2.dvec.double}
\title{Find Common Type for dvec, double}
\usage{
\method{vec_ptype2}{dvec.double}(x, y, ...)
}
\arguments{
\item{x}{dvec}

\item{y}{double}

\item{...}{ignored}
}
\description{
Find common type for dvec, double.
}
\examples{
str(vctrs::vec_ptype2(as_dvec(1), 1))
str(vctrs::vec_ptype2(1, as_dvec(1)))

str(vctrs::vec_ptype2(as_dvec(1, label = 'x'), 1))
str(vctrs::vec_ptype2(1, as_dvec(1, label= 'x')))
str(vctrs::vec_ptype2(as_dvec(1), structure(1, label = 'x')))

}
\keyword{internal}
