% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_ptype2.R, R/vec_ptype2_dvec.R
\name{vec_ptype2.dvec.dvec}
\alias{vec_ptype2.dvec.dvec}
\title{Find Common Type for dvec, dvec}
\usage{
\method{vec_ptype2}{dvec.dvec}(x, y, ...)

\method{vec_ptype2}{dvec.dvec}(x, y, ...)
}
\arguments{
\item{x}{dvec}

\item{y}{dvec}

\item{...}{ignored}
}
\description{
Find common type for dvec, dvec.

Find common type for dvec, dvec.
}
\examples{
str(vctrs::vec_ptype2(as_dvec(1L), as_dvec(1)))
str(vctrs::vec_ptype2(as_dvec(1), as_dvec(1L)))
str(vctrs::vec_ptype2(as_dvec(TRUE), as_dvec(1L)))
str(vctrs::vec_ptype2(as_dvec(TRUE), as_dvec(1)))
str(vctrs::vec_ptype2(as_dvec(1), as_dvec(1+0i)))

vctrs::vec_ptype2(
  structure(as_dvec(1), guide = 'mg'),
  structure(as_dvec(1), guide = 'kg')
)
str(vctrs::vec_ptype2(as_dvec(1L), as_dvec(1)))
str(vctrs::vec_ptype2(as_dvec(1), as_dvec(1L)))
str(vctrs::vec_ptype2(as_dvec(TRUE), as_dvec(1L)))
str(vctrs::vec_ptype2(as_dvec(TRUE), as_dvec(1)))
str(vctrs::vec_ptype2(as_dvec(1), as_dvec(1+0i)))

vctrs::vec_ptype2(
  structure(as_dvec(1), guide = 'mg'),
  structure(as_dvec(1), guide = 'kg')
)
}
\keyword{internal}
