% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_ptype2.R
\name{vec_ptype2.dvec.dvec}
\alias{vec_ptype2.dvec.dvec}
\title{Find Common Type for dvec, dvec}
\usage{
\method{vec_ptype2}{dvec.dvec}(x, y, ...)
}
\arguments{
\item{x}{dvec}

\item{y}{dvec}

\item{...}{passed arguments}
}
\description{
Find common type for dvec, dvec.
}
\examples{
vctrs::vec_ptype2(as_dvec('bar'), as_dvec(1))
vctrs::vec_ptype2(as_dvec(1), as_dvec('foo'))
vctrs::vec_ptype2(
  structure(as_dvec(1), guide = 'mg'),
  structure(as_dvec(1), guide = 'kg')
)
}
\keyword{internal}
