% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_ptype2.R
\name{vec_ptype2.dvec.logical}
\alias{vec_ptype2.dvec.logical}
\title{Find Common Type for dvec, logical}
\usage{
\method{vec_ptype2}{dvec.logical}(x, y, ...)
}
\arguments{
\item{x}{dvec}

\item{y}{logical}

\item{...}{ignored}
}
\description{
Find common type for dvec, logical.
}
\keyword{internal}
