% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_ptype2.R, R/vec_ptype2_dvec.R
\name{vec_ptype2.logical.dvec}
\alias{vec_ptype2.logical.dvec}
\title{Find Common Type for logical, dvec}
\usage{
\method{vec_ptype2}{logical.dvec}(x, y, ...)

\method{vec_ptype2}{logical.dvec}(x, y, ...)
}
\arguments{
\item{x}{logical}

\item{y}{dvec}

\item{...}{ignored}
}
\description{
Find common type for logical, dvec.

Find common type for logical, dvec.
}
\keyword{internal}
