% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-to-plotmath.R
\name{wiki_to_plotmath}
\alias{wiki_to_plotmath}
\title{Convert One Wiki Symbol to Plotmath}
\usage{
wiki_to_plotmath(
  x,
  unrecognized = getOption("plotmath_unrecognized", "quoteToken"),
  ...
)
}
\arguments{
\item{x}{character}

\item{unrecognized}{function to process unrecognized tokens: default \code{\link{quoteToken}}}

\item{...}{ignored}
}
\value{
character
}
\description{
Converts one wiki symbol to plotmath.
See description for \code{\link{as_wikisymbol}}.
By default, unrecognized tokens are returned
unmodified if they are parseable.
Otherwise, backslashes and single quotesare escaped,
and the result is wrapped in single quotes.
See \code{\link{quoteToken}}.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
lapply(x, wiki_to_plotmath)
library(magrittr)
render <- . \%>\% as_wikisymbol \%>\% as_plotmath \%>\% as.expression
'^*' \%>\% render
'^\\\\*' \%>\% render
'^\\\\*.' \%>\% render
'^\\\\.' \%>\% render
'^\\\\\\\\' \%>\% render
'\\\\\\\\' \%>\% render
'^\\\\^' \%>\% render
'^\\\\_' \%>\% render
'\\\\^' \%>\% render
'^.a' \%>\% render
'\\\\$' \%>\% render
wiki_to_plotmath('\\\\$') \%>\% goodToken
wiki_to_plotmath('\\\\$', unescape = FALSE) \%>\% goodToken
wiki_to_plotmath('\\\\$', unrecognized = function(x,...)x) \%>\% goodToken
options(plotmath_unrecognized = function(x,...)x)
wiki_to_plotmath('\\\\$') \%>\% goodToken
options(plotmath_unrecognized = NULL)
wiki_to_plotmath('\\\\$') \%>\% goodToken

}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol.symbolic_units}()},
\code{\link{as_wikisymbol.unit_string}()},
\code{\link{as_wikisymbol.units}()},
\code{\link{as_wikisymbol}()},
\code{\link{wiki_to_latex}()},
\code{\link{wikitoken}()}
}
\concept{wikisymbol}
\keyword{internal}
