% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{wikisym2latex_}
\alias{wikisym2latex_}
\title{Convert One Wiki Symbol to Latex}
\usage{
wikisym2latex_(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored}
}
\value{
character
}
\description{
Converts one wiki symbol to latex.
}
\details{
A Wiki symbol is simple text with arbitrarily
nested subscript (\code{_}) and superscript
(\code{^}) groupings.  Use dot (\code{.})
to explicitly terminate a grouping, and use
backslash-dot (\code{\.}) for a literal dot.
Examples: \code{V_c./F}. Trailing dots need
not be supplied. Leading/trailing whitespace
is removed. Tab character not allowed.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_wikisymbol(x)
lapply(x, wikisym2latex_)
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol}()},
\code{\link{wikisym2plotmath_}()}
}
\concept{wikisymbol}
\keyword{internal}
