% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikisymbol.R
\name{wikisym2plotmath_}
\alias{wikisym2plotmath_}
\title{Convert One Wiki Symbol to Plotmath}
\usage{
wikisym2plotmath_(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored}
}
\value{
character
}
\description{
Converts one wiki symbol to plotmath.
See description for \code{\link{as_wikisymbol}}.
Space characters are converted to plotmath space
(\code{phantom(0)}). Asterix is converted to \code{\%.\%}.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_wikisymbol(x)
lapply(x, wikisym2plotmath_)
}
\seealso{
Other wikisymbol: 
\code{\link{[.latex}()},
\code{\link{[.plotmath}()},
\code{\link{[.wikisymbol}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.wikisymbol}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.wikisymbol}()},
\code{\link{as_latex}()},
\code{\link{as_plotmath.wikisymbol}()},
\code{\link{as_plotmath}()},
\code{\link{as_wikisymbol.character}()},
\code{\link{as_wikisymbol.factor}()},
\code{\link{as_wikisymbol.symbolic_units}()},
\code{\link{as_wikisymbol}()},
\code{\link{wikisym2latex_}()}
}
\concept{wikisymbol}
\keyword{internal}
