% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlet.R
\name{write_yamlet}
\alias{write_yamlet}
\title{Write Yamlet}
\usage{
write_yamlet(
  x,
  con = stdout(),
  eol = "\\n",
  useBytes = FALSE,
  default_keys = getOption("yamlet_default_keys", list("label", "guide")),
  fileEncoding = getOption("encoding"),
  ...
)
}
\arguments{
\item{x}{something that can be coerced to class 'yamlet', like a yamlet object or a decorated data.frame}

\item{con}{passed to \code{\link{writeLines}}}

\item{eol}{end-of-line; passed to \code{\link{writeLines}} as \code{sep}}

\item{useBytes}{passed to \code{\link{writeLines}}}

\item{default_keys}{character: default keys for the first n anonymous members of each element}

\item{fileEncoding}{if \code{con} is character, passed to \code{\link{file}} as \code{encoding}}

\item{...}{passed to \code{\link{as_yamlet}}}
}
\value{
invisible character representation of yamlet (storage syntax)
}
\description{
Writes yamlet to file. Similar to \code{\link{io_yamlet.data.frame}}
but returns invisible storage format instead of invisible storage location.
}
\examples{
library(csv)
file <- system.file(package = 'yamlet', 'extdata','quinidine.csv')
meta <- system.file(package = 'yamlet', 'extdata','quinidine.yaml')
x <- as.csv(file)
y <- read_yamlet(meta)
x <- decorate(x, meta = y)
identical(x, decorate(file))
tmp <- tempfile()
write_yamlet(x, tmp)
identical(read_yamlet(meta), read_yamlet(tmp))

}
\seealso{
\code{\link{decorate.list}}

Other interface: 
\code{\link{decorate.character}()},
\code{\link{decorate.list}()},
\code{\link{decorate}()},
\code{\link{io_csv.character}()},
\code{\link{io_csv.data.frame}()},
\code{\link{io_csv}()},
\code{\link{io_table.character}()},
\code{\link{io_table.data.frame}()},
\code{\link{io_table}()},
\code{\link{io_yamlet.character}()},
\code{\link{io_yamlet.data.frame}()},
\code{\link{io_yamlet.yamlet}()},
\code{\link{io_yamlet}()},
\code{\link{read_yamlet}()}
}
\concept{interface}
